@@ -1,5 +1,5 @@
 {
-    "version": "4.0.5+20200926.2237",
+    "version": "4.0.6+20201103.2218",
     "license": "CC-BY-NC-SA-4.0",
     "scripts": {
         "predev": "tsc -b scripts && node scripts predev",
@@ -16,20 +16,20 @@
         "tsc:docs": "tsc -b scripts && node scripts tscDocs"
     },
     "dependencies": {
-        "@fortawesome/fontawesome-svg-core": "^1.2.30",
-        "@fortawesome/free-solid-svg-icons": "^5.14.0",
+        "@fortawesome/fontawesome-svg-core": "^1.2.32",
+        "@fortawesome/free-solid-svg-icons": "^5.15.1",
         "@fortawesome/vue-fontawesome": "^2.0.0",
-        "broadcast-channel": "^3.2.0",
+        "broadcast-channel": "^3.3.0",
         "combokeys": "^3.0.1",
-        "highcharts": "^8.1.2",
+        "highcharts": "^8.2.2",
         "localforage": "^1.8.1",
         "lodash": ">=4.17.19",
         "mini-svg-data-uri": "^1.2.3",
-        "moment": "^2.29.0",
         "nearley": "^2.19.7",
+        "moment": "^2.29.1",
         "ua-parser-js": "^0.7.22",
-        "vue": "<2.6.12",
-        "vue-i18n": "^8.21.1",
+        "vue": "^2.6.12",
+        "vue-i18n": "^8.22.1",
         "vue-js-modal": "2.0.0-rc.3",
         "vue-js-toggle-button": "^1.3.3",
         "vue-notification": "^1.3.20",
@@ -41,45 +41,43 @@
     "devDependencies": {
         "@types/combokeys": "^2.4.6",
         "@types/i18n-js": "^3.0.3",
-        "@types/jquery": "^3.5.1",
-        "@types/leaflet": "^1.5.17",
-        "@types/lodash": "^4.14.161",
+        "@types/jquery": "^3.5.4",
+        "@types/leaflet": "^1.5.19",
+        "@types/lodash": "^4.14.165",
         "@types/nearley": "^2.11.1",
-        "@types/node": "^14.11.2",
+        "@types/node": "^14.14.7",
         "@types/ua-parser-js": "^0.7.33",
-        "@types/vue": "^2.0.0",
-        "@types/vue-select": "^2.5.0",
+        "@types/vue-select": "^2.5.1",
         "@types/webpack-env": "^1.15.3",
-        "@typescript-eslint/eslint-plugin": "^3.10.1",
-        "@typescript-eslint/parser": "^3.10.1",
-        "@vuepress/plugin-active-header-links": "^1.5.4",
-        "@vuepress/plugin-back-to-top": "^1.5.4",
-        "all-contributors-cli": "^6.17.4",
+        "@typescript-eslint/eslint-plugin": "^4.7.0",
+        "@typescript-eslint/parser": "^4.7.0",
+        "@vuepress/plugin-active-header-links": "^1.7.0",
+        "@vuepress/plugin-back-to-top": "^1.7.0",
+        "all-contributors-cli": "^6.19.0",
         "cache-loader": "^4.1.0",
         "clean-webpack-plugin": "^3.0.0",
         "copy-dir": "^1.3.0",
         "css-loader": "^3.6.0",
-        "eslint": "^7.9.0",
+        "eslint": "^7.13.0",
         "eslint-plugin-prettier": "^3.1.4",
-        "eslint-plugin-vue": "^6.2.2",
-        "html-loader": "^1.3.1",
+        "eslint-plugin-vue": "^7.1.0",
+        "html-loader": "^1.3.2",
         "nearley-loader": "^2.0.0",
-        "sass": "^1.26.11",
-        "sass-loader": "^10.0.1",
+        "sass": "^1.29.0",
+        "sass-loader": "^10.1.0",
         "string-replace-loader": "^2.3.0",
-        "terser": "^5.3.2",
-        "ts-loader": "^8.0.4",
-        "tslint": "^6.1.2",
-        "typescript": "^4.0.3",
-        "url-loader": "^4.1.0",
-        "vue-loader": "^15.9.3",
-        "vue-template-compiler": "<2.6.12",
-        "vuepress": "^1.5.4",
+        "terser": "^5.3.8",
+        "ts-loader": "^8.0.11",
+        "typescript": "^4.0.5",
+        "url-loader": "^4.1.1",
+        "vue-loader": "^15.9.5",
+        "vue-template-compiler": "^2.6.12",
+        "vuepress": "^1.7.0",
         "vuepress-plugin-redirect": "^1.2.5",
         "vuepress-plugin-smooth-scroll": "^0.0.10",
         "vuepress-plugin-zooming": "^1.1.8",
         "webpack": "^4.44.2",
-        "webpack-cli": "^3.3.12",
+        "webpack-cli": "^4.2.0",
         "webpack-loader-append-prepend": "^1.0.2"
     },
     "engines": {
