@@ -1,5 +1,4 @@
 import packageJson from '../package.json';
-import staticConfigs from '../static/.configs.json';
 import moment from 'moment';
 import fs from 'fs';
 
@@ -8,9 +7,9 @@ packageJson.version = packageJson.version.replace(
     `+${moment().format('YYYYMMDD.HHmm')}`
 );
 
-// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-ignore
-if (!staticConfigs.hasOwnProperty('versions')) staticConfigs.versions = {};
+const staticConfigs = {} as Record<string, unknown>;
+
+staticConfigs.versions = {};
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
 staticConfigs.versions[process.argv[2] === 'production' ? 'stable' : 'beta'] =
