@@ -8,9 +8,9 @@ const sortJSON = (object: unknown, arrays = false): unknown => {
     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
     // @ts-ignore
     return Object.fromEntries(
-        Object.entries(object).sort(([keyA], [keyB]) =>
-            keyA < keyB ? -1 : keyA > keyB ? 1 : 0
-        )
+        Object.entries(object)
+            .sort(([keyA], [keyB]) => (keyA < keyB ? -1 : keyA > keyB ? 1 : 0))
+            .map(([key, value]) => [key, sortJSON(value)])
     );
 };
 
