@@ -4,24 +4,29 @@ export default {
     admins: ['de_DE-205976', 'de_DE-675'],
     browser: {
         chrome: {
-            latest: 84,
+            latest: 86,
             supported: 84,
             download: 'https://chrome.com',
         },
         firefox: {
-            latest: 79,
+            latest: 82,
             supported: 79,
             download: 'https://firefox.com',
         },
         opera: {
-            latest: 70,
+            latest: 72,
             supported: 70,
             download: 'https://opera.com',
         },
         safari: {
-            latest: 13,
+            latest: 14,
             supported: 13,
             download: 'https://apple.com/safari/',
+        },
+        edge: {
+            latest: 86,
+            supported: 86,
+            download: 'https://microsoft.com/edge',
         },
     },
     discord: 'https://discord.gg/RcTNjpB',
@@ -30,108 +35,103 @@ export default {
     games: {
         cs_CZ: {
             flag: '🇨🇿',
-            locale_fallback: 'en_US',
             name: 'Operacni-Stredisko.cz',
             shortURL: 'operacni-stredisko.cz',
+            police: 'policie',
         },
         da_DK: {
             flag: '🇩🇰',
-            locale_fallback: 'en_US',
             name: 'Alarmcentral-Spil.dk',
             shortURL: 'alarmcentral-spil.dk',
+            police: 'politi',
         },
         de_DE: {
             flag: '🇩🇪',
             name: 'Leitstellenspiel.de',
             shortURL: 'leitstellenspiel.de',
+            police: 'polizei',
         },
-        el_GR: {
-            flag: '🇬🇷',
-            locale_fallback: 'en_US',
-            name: 'Missionchief Greece',
-            shortURL: 'missionchief.gr',
-        },
+        // el_GR: {
+        //     flag: '🇬🇷',
+        //     name: 'Missionchief Greece',
+        //     shortURL: 'missionchief.gr',
+        // },
         en_AU: {
             flag: '🇦🇺',
-            locale_fallback: 'en_GB',
             name: 'Missionchief-Australia.com',
             shortURL: 'missionchief-australia.com',
+            police: 'police',
         },
         en_GB: {
             flag: '🇬🇧',
-            locale_fallback: 'en_US',
             name: 'Missionchief.co.uk',
             shortURL: 'missionchief.co.uk',
+            police: 'police',
         },
         en_US: {
             flag: '🇺🇸',
-            locale_fallback: 'de_DE',
             name: 'Missionschief.com',
             shortURL: 'missionchief.com',
+            police: 'police',
         },
         es_ES: {
             flag: '🇪🇸',
-            locale_fallback: 'en_US',
             name: 'Centro-de-Mando.es',
             shortURL: 'centro-de-mando.es',
         },
         es_MX: {
             flag: '🇲🇽',
-            locale_fallback: 'es_ES',
             name: 'Centro-de-Mando.mx',
             shortURL: 'centro-de-mando.mx',
         },
         fi_FI: {
             flag: '🇫🇮',
-            locale_fallback: 'en_US',
             name: 'Hätäkeskuspeli.com',
             shortURL: 'hatakeskuspeli.com',
+            police: 'poliisi',
         },
         fr_FR: {
             flag: '🇫🇷',
-            locale_fallback: 'en_US',
             name: 'Operateur112.fr',
             shortURL: 'operateur112.fr',
+            police: 'police',
         },
         it_IT: {
             flag: '🇮🇹',
-            locale_fallback: 'en_US',
             name: 'Operatore112.it',
             shortURL: 'operatore112.it',
+            police: 'polizia',
         },
         ja_JP: {
             flag: '🇯🇵',
-            locale_fallback: 'en_US',
             name: 'Missionchief-Japan.com',
             shortURL: 'missionchief-japan.com',
         },
         ko_KR: {
             flag: '🇰🇷',
-            locale_fallback: 'en_US',
             name: 'Missionchief-Korea.com',
             shortURL: 'missionchief-korea.com',
         },
         nb_NO: {
             flag: '🇳🇴',
-            locale_fallback: 'en_US',
             name: 'Nodsentralspillet.com',
             shortURL: 'nodsentralspillet.com',
+            police: 'politiet',
         },
         nl_NL: {
             flag: '🇳🇱',
-            locale_fallback: 'en_GB',
             name: 'Meldkamerspel.com',
             shortURL: 'meldkamerspel.com',
+            police: 'politie',
         },
         pl_PL: {
             flag: '🇵🇱',
-            locale_fallback: 'en_US',
             name: 'Operatorraunkowy.pl',
             shortURL: 'operatorratunkowy.pl',
+            police: 'policja',
         },
         pt_BR: {
             flag: '🇧🇷',
-            locale_fallback: 'pt_PT',
             name: 'Operador193.com',
             shortURL: 'operador193.com',
         },
@@ -140,16 +140,15 @@ export default {
             locale_fallback: 'en_US',
             name: 'Jogo-operador112.com',
             shortURL: 'jogo-operador112.com',
+            police: 'policia',
         },
         ro_RO: {
             flag: '🇷🇴',
-            locale_fallback: 'en_US',
             name: 'Jocdispecerat112.com',
             shortURL: 'jocdispecerat112.com',
         },
         ru_RU: {
             flag: '🇷🇺',
-            locale_fallback: 'en_US',
             name: 'Dispetcher112.ru',
             shortURL: 'dispetcher112.ru',
         },
@@ -161,19 +160,17 @@ export default {
         },
         sv_SE: {
             flag: '🇸🇪',
-            locale_fallback: 'en_US',
             name: 'Larmcentralen-spelet.se',
             shortURL: 'larmcentralen-spelet.se',
+            police: 'polis',
         },
         tr_TR: {
             flag: '🇹🇷',
-            locale_fallback: 'en_US',
             name: '112-Merkez.com',
             shortURL: '112-merkez.com',
         },
         uk_UA: {
             flag: '🇺🇦',
-            locale_fallback: 'en_US',
             name: 'Dyspetcher101-game.com',
             shortURL: 'dyspetcher101-game.com',
         },
@@ -185,5 +182,5 @@ export default {
         'core-modules': ['telemetry', 'releasenotes', 'support'],
     },
     prefix: 'lssmv4',
-    server: 'https://v4.lssm.ledbrain.de/',
+    server: 'https://proxy.lss-manager.de/v4/',
 } as Config;
