@@ -30,7 +30,14 @@ const modules = {
                 'Möchtest du wirklich Einstellungen auf ihre Standardwerte zurücksetzen? Dies lässt sich nicht rückgängig machen!',
             close: 'Abbrechen',
             total: 'Alle Einstellungen',
-            module: 'Nur von diesem Modul',
+            module: 'Nur von diesem Modul <b>{module}</b>',
+        },
+        resetWarningSetting: {
+            title: 'Einstellung zurücksetzen',
+            text:
+                'Möchtest du wirklich diese eine Einstellung <b>{setting}</b> des Moduls <b>{module}</b> auf ihren Standartwert zurücksetzen?',
+            close: 'Abbrechen',
+            reset: 'Zurücksetzen',
         },
         closeWarning: {
             title: 'Ungespeicherte Änderungen',
@@ -81,914 +88,1022 @@ export default {
                 'Zeigt in der Navigationsleiste statt dem LSSM-Logo ein einfaches Label an.',
         },
     },
-    vehicles: [
-        {
+    vehicles: {
+        0: {
             caption: 'LF 20',
             color: '#cc0000',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 9,
-            wtank: 2_000,
+            wtank: 2000,
+            icon: 'truck-moving',
         },
-        {
+        1: {
             caption: 'LF 10',
             color: '#bb0000',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 9,
-            wtank: 1_200,
+            wtank: 1200,
+            icon: 'truck-moving',
         },
-        {
+        2: {
             caption: 'DLK 23',
             color: '#d92626',
             coins: 30,
-            credits: 10_000,
+            credits: 10000,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 0,
             special: 'Wird ab 3 Feuerwachen benötigt.',
+            icon: 'truck',
         },
-        {
+        3: {
             caption: 'ELW 1',
             color: '#d02525',
             coins: 25,
-            credits: 10_000,
+            credits: 10000,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 0,
             special: 'Wird ab 6 Feuerwachen benötigt.',
+            icon: 'car-side',
         },
-        {
+        4: {
             caption: 'RW',
             color: '#ad1f1f',
             coins: 25,
-            credits: 12_180,
+            credits: 12180,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 0,
             special: 'Wird ab 4 Feuerwachen benötigt.',
+            icon: 'truck',
         },
-        {
+        5: {
             caption: 'GW-A',
             color: '#9c1c1c',
             coins: 25,
-            credits: 11_680,
+            credits: 11680,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 0,
             special: 'Wird ab 5 Feuerwachen benötigt.',
+            icon: 'truck',
         },
-        {
+        6: {
             caption: 'LF 8/6',
             color: '#aa0000',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 9,
             wtank: 600,
+            icon: 'truck-moving',
         },
-        {
+        7: {
             caption: 'LF 20/16',
             color: '#990000',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 9,
-            wtank: 1_600,
+            wtank: 1600,
+            icon: 'truck-moving',
         },
-        {
+        8: {
             caption: 'LF 10/6',
             color: '#880000',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 9,
             wtank: 600,
+            icon: 'truck-moving',
         },
-        {
+        9: {
             caption: 'LF 16-TS',
             color: '#770000',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 9,
             wtank: 0,
+            icon: 'truck-moving',
         },
-        {
+        10: {
             caption: 'GW-Öl',
             color: '#8b1818',
             coins: 25,
-            credits: 12_000,
+            credits: 12000,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 0,
             special: 'Wird ab 6 Feuerwachen benötigt.',
+            icon: 'truck',
         },
-        {
+        11: {
             caption: 'GW-L2-Wasser',
             color: '#e61919',
             coins: 25,
-            credits: 17_300,
+            credits: 17300,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 0,
             special:
                 'Ein Schlauchwagen wird ab 7 Feuerwachen benötigt. Er gibt 25% Bonus auf die Wassermenge, die durch (H/T)LF vor Ort ist.',
+            icon: 'truck',
         },
-        {
+        12: {
             caption: 'GW-Messtechnik',
             color: '#791515',
             coins: 25,
-            credits: 18_300,
+            credits: 18300,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 0,
             schooling: 'Feuerwehr - GW-Messtechnik Lehrgang',
             shownSchooling: 'GW-Messtechnik',
             special: 'Wird ab 10 Feuerwachen benötigt.',
+            icon: 'truck',
         },
-        {
+        13: {
             caption: 'SW 1000',
             color: '#dc1818',
             coins: 25,
-            credits: 17_300,
+            credits: 17300,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 0,
             special:
                 'Ein Schlauchwagen wird ab 7 Feuerwachen benötigt. Er gibt 25% Bonus auf die Wassermenge, die durch (H/T)LF vor Ort ist.',
+            icon: 'truck',
         },
-        {
+        14: {
             caption: 'SW 2000',
             color: '#ca1616',
             coins: 25,
-            credits: 17_300,
+            credits: 17300,
             minPersonnel: 1,
             maxPersonnel: 6,
             wtank: 0,
             special:
                 'Ein Schlauchwagen wird ab 7 Feuerwachen benötigt. Er gibt 25% Bonus auf die Wassermenge, die durch (H/T)LF vor Ort ist.',
+            icon: 'truck',
         },
-        {
+        15: {
             caption: 'SW 2000-Tr',
             color: '#b81414',
             coins: 25,
-            credits: 17_300,
+            credits: 17300,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 0,
             special:
                 'Ein Schlauchwagen wird ab 7 Feuerwachen benötigt. Er gibt 25% Bonus auf die Wassermenge, die durch (H/T)LF vor Ort ist.',
+            icon: 'truck',
         },
-        {
+        16: {
             caption: 'SW Kats',
             color: '#a51212',
             coins: 25,
-            credits: 17_300,
+            credits: 17300,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 0,
             special:
                 'Ein Schlauchwagen wird ab 7 Feuerwachen benötigt. Er gibt 25% Bonus auf die Wassermenge, die durch (H/T)LF vor Ort ist.',
+            icon: 'truck',
         },
-        {
+        17: {
             caption: 'TLF 2000',
             color: '#cc2222',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 3,
-            wtank: 2_000,
+            wtank: 2000,
+            icon: 'truck-moving',
         },
-        {
+        18: {
             caption: 'TLF 3000',
             color: '#bb2222',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 3,
-            wtank: 3_000,
+            wtank: 3000,
+            icon: 'truck-moving',
         },
-        {
+        19: {
             caption: 'TLF 8/8',
             color: '#aa2222',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 800,
+            icon: 'truck-moving',
         },
-        {
+        20: {
             caption: 'TLF 8/18',
             color: '#992222',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 1800,
+            icon: 'truck-moving',
         },
-        {
+        21: {
             caption: 'TLF 16/24-Tr',
             color: '#882222',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 2400,
+            icon: 'truck-moving',
         },
-        {
+        22: {
             caption: 'TLF 16/25',
             color: '#772222',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 6,
             wtank: 2400,
+            icon: 'truck-moving',
         },
-        {
+        23: {
             caption: 'TLF 16/45',
             color: '#662222',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 4500,
+            icon: 'truck-moving',
         },
-        {
+        24: {
             caption: 'TLF 20/40',
             color: '#552222',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 4000,
+            icon: 'truck-moving',
         },
-        {
+        25: {
             caption: 'TLF 20/40-SL',
             color: '#442222',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 4000,
+            icon: 'truck-moving',
         },
-        {
+        26: {
             caption: 'TLF 16',
             color: '#332222',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 1800,
+            icon: 'truck-moving',
         },
-        {
+        27: {
             caption: 'GW-Gefahrgut',
             color: '#681212',
             coins: 25,
-            credits: 19_200,
+            credits: 19200,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 0,
             schooling: 'Feuerwehr - GW-Gefahrgut Lehrgang',
             shownSchooling: 'GW-Gefahrgut',
             special: 'Wird ab 11 Feuerwachen benötigt.',
+            icon: 'truck',
         },
-        {
+        28: {
             caption: 'RTW',
             color: '#f59f00',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 2,
+            icon: 'ambulance',
         },
-        {
+        29: {
             caption: 'NEF',
             color: '#e09200',
             coins: 25,
-            credits: 20_000,
+            credits: 20000,
             minPersonnel: 1,
             maxPersonnel: 2,
             schooling: 'Rettungsdienst - Notarzt-Ausbildung',
             shownSchooling: 'Notarzt',
             special: 'Wird ab 3 Rettungswachen benötigt.',
+            icon: 'car_side',
         },
-        {
+        30: {
             caption: 'HLF 20',
             color: '#440000',
             coins: 25,
-            credits: 20_000,
+            credits: 20000,
             minPersonnel: 1,
             maxPersonnel: 9,
             wtank: 1600,
+            icon: 'truck-moving',
         },
-        {
+        31: {
             caption: 'RTH',
             color: '#ddaa3c',
             coins: 30,
-            credits: 300_000,
+            credits: 300000,
             minPersonnel: 1,
             maxPersonnel: 1,
             schooling: 'Rettungsdienst - Notarzt-Ausbildung',
             shownSchooling: 'Notarzt',
             special: 'Wird ab 3 Rettungswachen benötigt.',
+            icon: 'helicopter',
         },
-        {
+        32: {
             caption: 'FuStW',
             color: '#005500',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 2,
+            icon: 'hamburger',
         },
-        {
+        33: {
             caption: 'GW-Höhenrettung',
             color: '#570f0f',
             coins: 25,
-            credits: 19_500,
+            credits: 19500,
             minPersonnel: 1,
             maxPersonnel: 9,
             wtank: 0,
             schooling: 'Feuerwehr - Höhenrettung Lehrgang',
             shownSchooling: 'GW-Höhenrettung',
             special: 'Wird ab 12 Feuerwachen benötigt.',
+            icon: 'truck',
         },
-        {
+        34: {
             caption: 'ELW 2',
             color: '#bf2222',
             coins: 25,
-            credits: 25_500,
+            credits: 25500,
             minPersonnel: 1,
             maxPersonnel: 6,
             wtank: 0,
             schooling: 'Feuerwehr - ELW 2 Lehrgang',
             shownSchooling: 'ELW 2',
             special: 'Wird ab 13 Feuerwachen benötigt.',
+            icon: 'truck',
         },
-        {
+        35: {
             caption: 'leBefKw',
             color: '#288f28',
             coins: 25,
-            credits: 10_000,
+            credits: 10000,
             minPersonnel: 1,
             maxPersonnel: 3,
             schooling: 'Polizei - Zugführer (leBefKw)',
             shownSchooling: 'Zugführer (leBefKw)',
+            icon: 'bus',
         },
-        {
+        36: {
             caption: 'MTW',
             color: '#450c0c',
             coins: 12,
-            credits: 2_500,
+            credits: 2500,
             minPersonnel: 1,
             maxPersonnel: 9,
             wtank: 0,
+            icon: 'shuttle-van',
         },
-        {
+        37: {
             caption: 'TSF-W',
             color: '#220000',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 6,
             wtank: 500,
+            icon: 'truck-moving',
         },
-        {
+        38: {
             caption: 'KTW',
             color: '#ffb61a',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 2,
+            icon: 'ambulance',
         },
-        {
+        39: {
             caption: 'GKW',
             color: '#00138f',
             coins: 25,
-            credits: 13_000,
+            credits: 13000,
             minPersonnel: 1,
             maxPersonnel: 9,
+            icon: 'truck-moving',
         },
-        {
+        40: {
             caption: 'MTW-TZ',
             color: '#0016a3',
             coins: 12,
-            credits: 2_500,
+            credits: 2500,
             minPersonnel: 1,
             maxPersonnel: 4,
             schooling: 'THW - Zugtrupp',
             shownSchooling: 'Zugtrupp',
+            icon: 'shuttle-van',
         },
-        {
+        41: {
             caption: 'MzKW',
             color: '#0018b8',
             coins: 25,
-            credits: 15_000,
+            credits: 15000,
             minPersonnel: 1,
             maxPersonnel: 4,
+            icon: 'truck-moving',
         },
-        {
+        42: {
             caption: 'LKW K 9',
             color: '#001bcc',
             coins: 25,
-            credits: 15_000,
+            credits: 15000,
             minPersonnel: 1,
             maxPersonnel: 3,
             schooling: 'THW - Fachgruppe Räumen',
             shownSchooling: 'Fachgruppe Räumen',
+            icon: 'truck-pickup',
         },
-        {
+        43: {
             caption: 'BRmG R',
             color: '#001ee0',
             coins: 25,
-            credits: 15_000,
+            credits: 15000,
             special: 'Muss vom "LKW K 9" zum Einsatz gezogen werden',
+            icon: 'snowplow',
         },
-        {
+        44: {
             caption: 'Anh DLE',
             color: '#0021f5',
             coins: 25,
-            credits: 15_000,
+            credits: 15000,
             special:
                 'Muss von einem "GKW", "MzKW", "MTW-TZ" oder "MLW 5" zum Einsatz gezogen werden.',
+            icon: 'trailer',
         },
-        {
+        45: {
             caption: 'MLW 5',
             color: '#0a2bff',
             coins: 12,
-            credits: 2_500,
+            credits: 2500,
             minPersonnel: 1,
             maxPersonnel: 6,
             schooling: 'THW - Fachgruppe Räumen',
             shownSchooling: 'Fachgruppe Räumen',
+            icon: 'truck-moving',
         },
-        {
+        46: {
             caption: 'WLF',
             color: '#f04242',
             coins: 12,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 0,
             schooling: 'Feuerwehr - Wechsellader Lehrgang',
             shownSchooling: 'Wechsellader',
+            truck: 'truck-pickup',
         },
-        {
+        47: {
             caption: 'AB-Rüst',
             color: '#ad0e0e',
             coins: 12,
-            credits: 6_000,
+            credits: 6000,
+            icon: 'square',
         },
-        {
+        48: {
             caption: 'AB-Atemschutz',
             color: '#9c0b0b',
             coins: 12,
-            credits: 6_000,
+            credits: 6000,
+            icon: 'square',
         },
-        {
+        49: {
             caption: 'AB-Öl',
             color: '#8b0707',
             coins: 12,
-            credits: 6_000,
+            credits: 6000,
+            icon: 'square',
         },
-        {
+        50: {
             caption: 'GruKw',
             color: '#2d9f29',
             coins: 25,
-            credits: 10_000,
+            credits: 10000,
             minPersonnel: 1,
             maxPersonnel: 9,
+            icon: 'bus',
         },
-        {
+        51: {
             caption: 'FüKw',
             color: '#31af31',
             coins: 25,
-            credits: 17_500,
+            credits: 17500,
             minPersonnel: 1,
             maxPersonnel: 3,
             schooling: 'Polizei - Hundertschaftsführer (FüKw)',
             shownSchooling: 'Hundertschaftsführer (FüKw)',
+            icon: 'car-side',
         },
-        {
+        52: {
             caption: 'GefKw',
             color: '#36bf36',
             coins: 25,
-            credits: 13_000,
+            credits: 13000,
             minPersonnel: 1,
             maxPersonnel: 3,
+            icon: 'bus-alt',
         },
-        {
+        53: {
             caption: 'Dekon-P',
             color: '#450909',
             coins: 25,
-            credits: 23_100,
+            credits: 23100,
             minPersonnel: 1,
             maxPersonnel: 6,
             schooling: 'Feuerwehr - Dekon-P Lehrgang',
             shownSchooling: 'Dekon-P',
             special:
                 'Pro Dekon-P müssen 6 Personen mit Dekon-P-Ausbildung vor Ort sein. Wird ab 14 Feuerwachen benötigt.',
+            icon: 'truck',
         },
-        {
+        54: {
             caption: 'AB-Dekon-P',
             color: '#450707',
             coins: 12,
-            credits: 6_000,
+            credits: 6000,
+            icon: 'square',
         },
-        {
+        55: {
             caption: 'KdoW-LNA',
             color: '#805c0f',
             coins: 25,
-            credits: 20_000,
+            credits: 20000,
             minPersonnel: 1,
             maxPersonnel: 1,
             wtank: 0,
             schooling: 'Rettungsdienst - LNA-Ausbildung',
             shownSchooling: 'LNA',
             special: 'Wird ab 5 Rettungswachen benötigt.',
+            icon: 'car-side',
         },
-        {
+        56: {
             caption: 'KdoW-OrgL',
             color: '#926911',
             coins: 25,
-            credits: 20_000,
+            credits: 20000,
             minPersonnel: 1,
             maxPersonnel: 1,
             wtank: 0,
             schooling: 'Rettungsdienst - OrgL-Ausbildung',
             shownSchooling: 'OrgL',
             special: 'Wird ab 10 Rettungswachen benötigt.',
+            icon: 'car-side',
         },
-        {
+        57: {
             caption: 'FwK',
             color: '#230606',
             coins: 25,
-            credits: 30_000,
+            credits: 30000,
             minPersonnel: 1,
             maxPersonnel: 2,
             wtank: 0,
             schooling: 'Feuerwehr - Feuerwehrkran Lehrgang',
             shownSchooling: 'Feuerwehrkran',
             special: 'Wird ab 14 Feuerwachen benötigt.',
+            icon: 'truck',
         },
-        {
+        58: {
             caption: 'KTW Typ B',
             color: '#cc5200',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 2,
             special:
                 'Kann, wenn ein GW-San vor Ort ist, behandelte Patienten vom Einsatzort abtransportieren. Ein Transport mit Notarztbegleitung durch ein NEF ist auch möglich.',
+            icon: 'ambulance',
         },
-        {
+        59: {
             caption: 'ELW 1 (SEG)',
             color: '#a34100',
             coins: 25,
-            credits: 25_500,
+            credits: 25500,
             minPersonnel: 1,
             maxPersonnel: 2,
             schooling: 'Rettungsdienst - SEG - Einsatzleitung',
             shownSchooling: 'Einsatzleitung (SEG)',
             special:
                 'Kann automatisch eingehende Sprechwünsche der eigenen Rettungsmittel bearbeiten und ihnen im Transportfall ein geeignetes Krankenhaus zuweisen. Die automatische Zuweisungsfunktion kann in dessen Einstellungen verändert werden.',
+            icon: 'car-side',
         },
-        {
+        60: {
             caption: 'GW-San',
             color: '#8f3900',
             coins: 25,
-            credits: 10_000,
+            credits: 10000,
             minPersonnel: 6,
             maxPersonnel: 6,
             schooling: 'Rettungsdienst - SEG - GW-San',
             shownSchooling: 'GW-San',
             special:
                 'Pro GW-San können 5 KTW Typ B wie ein RTW arbeiten. Er behandelt die Patienten zwar größtenteils, aber für die letzten paar Prozente braucht man dann einen RTW, KTW Typ B oder ähnliches.',
+            icon: 'truck',
         },
-        {
+        61: {
             caption: 'Polizeihubschrauber',
             color: '#14743f',
             coins: 30,
-            credits: 300_000,
+            credits: 300000,
             minPersonnel: 1,
             maxPersonnel: 3,
             schooling: 'Polizei - Polizeihubschrauber',
             shownSchooling: 'Polizeihubschrauber',
+            icon: 'helicopter',
         },
-        {
+        62: {
             caption: 'AB-Schlauch',
             color: '#e60808',
             coins: 12,
-            credits: 6_000,
+            credits: 6000,
+            icon: 'square',
         },
-        {
+        63: {
             caption: 'GW-Taucher',
             color: '#88ecc4',
             coins: 25,
-            credits: 10_000,
+            credits: 10000,
             minPersonnel: 2,
             maxPersonnel: 2,
             schooling: 'Rettungsdienst - GW-Taucher Lehrgang',
             shownSchooling: 'GW-Taucher',
+            icon: 'truck',
         },
-        {
+        64: {
             caption: 'GW-Wasserrettung',
             color: '#91dab5',
             coins: 25,
-            credits: 10_000,
+            credits: 10000,
             minPersonnel: 1,
             maxPersonnel: 6,
             schooling: 'Rettungsdienst - GW-Wasserrettung Lehrgang',
             shownSchooling: 'GW-Wasserrettung',
+            icon: 'truck',
         },
-        {
+        65: {
             caption: 'LKW 7 Lkr 19 tm',
             color: '#123183',
             coins: 25,
-            credits: 10_000,
+            credits: 10000,
             minPersonnel: 1,
             maxPersonnel: 2,
+            icon: 'truck-pickup',
         },
-        {
+        66: {
             caption: 'Anh MzB',
             color: '#1b428a',
             coins: 12,
-            credits: 6_000,
+            credits: 6000,
             special:
                 'Muss von einem "LKW 7 Lkr 19 tm" zum Einsatz gezogen werden. Pro gefordertem Boot müssen mind. 4 Personen mit der Ausbildung "GW-Wasserrettung" oder "Wassergefahren" vor Ort sein!',
+            icon: 'trailer',
         },
-        {
+        67: {
             caption: 'Anh SchlB',
             color: '#245390',
             coins: 12,
-            credits: 6_000,
+            credits: 6000,
             special:
                 'Muss von einem "LKW 7 Lkr 19 tm" zum Einsatz gezogen werden. Pro gefordertem Boot müssen mind. 4 Personen mit der Ausbildung "GW-Wasserrettung" oder "Wassergefahren" vor Ort sein!',
+            icon: 'trailer',
         },
-        {
+        68: {
             caption: 'Anh MzAB',
             color: '#296497',
             coins: 12,
-            credits: 6_000,
+            credits: 6000,
             special:
                 'Muss von einem "LKW 7 Lkr 19 tm" zum Einsatz gezogen werden. Pro gefordertem Boot müssen mind. 4 Personen mit der Ausbildung "GW-Wasserrettung" oder "Wassergefahren" vor Ort sein!',
+            icon: 'trailer',
         },
-        {
+        69: {
             caption: 'Tauchkraftwagen',
             color: '#36759e',
             coins: 25,
-            credits: 10_000,
+            credits: 10000,
             minPersonnel: 1,
             maxPersonnel: 2,
             schooling: 'THW - Bergungstaucher Lehrgang',
             shownSchooling: 'GW-Taucher',
+            icon: 'truck',
         },
-        {
+        70: {
             caption: 'MZB',
             color: '#9ac8a6',
             coins: 12,
-            credits: 6_000,
+            credits: 6000,
             special:
                 'Muss von einem "GW-Wasserrettung" oder einem "GW-Taucher" zum Einsatz gezogen werden. Pro gefordertem Boot müssen mind. 4 Personen mit der Ausbildung "GW-Wasserrettung" oder "Wassergefahren" vor Ort sein!',
+            icon: 'trailer',
         },
-        {
+        71: {
             caption: 'AB-MZB',
             color: '#bf7f6a',
             coins: 12,
-            credits: 6_000,
+            credits: 6000,
             special:
                 'Pro gefordertem Boot müssen mind. 4 Personen mit der Ausbildung "GW-Wasserrettung" oder "Wassergefahren" vor Ort sein!',
+            icon: 'square',
         },
-        {
+        72: {
             caption: 'WaWe 10',
             color: '#36b15d',
             coins: 25,
-            credits: 13_000,
+            credits: 13000,
             minPersonnel: 5,
             maxPersonnel: 5,
             schooling: 'Polizei - Wasserwerfer',
             shownSchooling: 'Wasserwerfer',
+            icon: 'bus-alt',
         },
-        {
+        73: {
             caption: 'GRTW',
             color: '#cc8500',
             coins: 25,
-            credits: 25_000,
+            credits: 25000,
             minPersonnel: 6,
             maxPersonnel: 6,
             special:
                 'Der GRTW kann in zwei verschiedenen Modi alarmiert werden:<br>1.Für leichtverletzte Personen, die keinen Notarzt benötigen. Dann können in dem GRTW 7 Patienten gleichzeitig behandelt und abtransportiert werden.<br>2.Für schwerverletzte Personen, die einen Notarzt benötigen. Dann können in dem GRTW 3 Patienten gleichzeitig behandelt werden.<br>Pro 20 Rettungswachen (bzw. 15 Rettungswachen mit Premium Account) kann ein GRTW gekauft werden.',
+            icon: 'ambulance',
         },
-        {
+        74: {
             caption: 'NAW',
             color: '#b87700',
             coins: 25,
-            credits: 25_000,
+            credits: 25000,
             minPersonnel: 3,
             maxPersonnel: 3,
             schooling: 'Rettungsdienst - Notarzt-Ausbildung',
             shownSchooling: 'Notarzt',
             special:
                 'Insgesamt so viele NAW kaufbar wie Rettungswachen (und auf einsatzbereit gestellte Rettungsdienst-Erweiterungen) vorhanden. Alle auch auf einer Wache patzierbar.',
+            icon: 'ambulance',
         },
-        {
+        75: {
             caption: 'FLF',
             color: '#7a534a',
             coins: 25,
-            credits: 80_000,
+            credits: 80000,
             minPersonnel: 2,
             maxPersonnel: 3,
             wtank: 12000,
             schooling: 'Feuerwehr - Flugfeldlöschfahrzeug-Ausbildung',
             shownSchooling: 'Flugfeldlöschfahrzeug',
+            icon: 'truck-moving',
         },
-        {
+        76: {
             caption: 'Rettungstreppe',
             color: '#6f6157',
             coins: 25,
-            credits: 20_000,
+            credits: 20000,
             minPersonnel: 2,
             maxPersonnel: 2,
             wtank: 0,
             schooling: 'Feuerwehr - Rettungstreppen-Ausbildung',
             shownSchooling: 'Rettungstreppe',
+            icon: 'truck',
         },
-        {
+        77: {
             caption: 'AB-Gefahrgut',
             color: '#680101',
             coins: 12,
-            credits: 6_000,
+            credits: 6000,
+            icon: 'square',
         },
-        {
+        78: {
             caption: 'AB-Einsatzleitung',
             color: '#bf1111',
             coins: 12,
-            credits: 6_000,
+            credits: 6000,
+            icon: 'square',
         },
-        {
+        79: {
             caption: 'SEK - ZF',
             color: '#217d1a',
             coins: 25,
-            credits: 10_000,
+            credits: 10000,
             minPersonnel: 3,
             maxPersonnel: 4,
             schooling: 'Polizei - SEK',
             shownSchooling: 'SEK',
+            icon: 'shuttle-van',
         },
-        {
+        80: {
             caption: 'SEK - MTF',
             color: '#1a7d1b',
             coins: 25,
-            credits: 10_000,
+            credits: 10000,
             minPersonnel: 9,
             maxPersonnel: 9,
             schooling: 'Polizei - SEK',
             shownSchooling: 'SEK',
+            icon: 'bus',
         },
-        {
+        81: {
             caption: 'MEK - ZF',
             color: '#1a7d22',
             coins: 25,
-            credits: 10_000,
+            credits: 10000,
             minPersonnel: 3,
             maxPersonnel: 4,
             schooling: 'Polizei - MEK',
             shownSchooling: 'MEK',
+            icon: 'shuttle-van',
         },
-        {
+        82: {
             caption: 'MEK - MTF',
             color: '#1a7d2a',
             coins: 25,
-            credits: 10_000,
+            credits: 10000,
             minPersonnel: 9,
             maxPersonnel: 9,
             schooling: 'Polizei - MEK',
             shownSchooling: 'MEK',
+            icon: 'bus',
         },
-        {
+        83: {
             caption: 'GW-Werkfeuerwehr',
             color: '#99222b',
             coins: 30,
-            credits: 15_000,
+            credits: 15000,
             minPersonnel: 1,
             maxPersonnel: 9,
             wtank: 0,
             schooling: 'Feuerwehr - Werkfeuerwehr-Ausbildung',
             shownSchooling: 'Werkfeuerwehr',
+            icon: 'truck',
         },
-        {
+        84: {
             caption: 'ULF mit Löscharm',
             color: '#992234',
             coins: 25,
-            credits: 20_000,
+            credits: 20000,
             minPersonnel: 1,
             maxPersonnel: 3,
-            wtank: 5_000,
+            wtank: 5000,
             schooling: 'Feuerwehr - Werkfeuerwehr-Ausbildung',
             shownSchooling: 'Werkfeuerwehr',
+            icon: 'truck-moving',
         },
-        {
+        85: {
             caption: 'TM 50',
             color: '#992b22',
             coins: 30,
-            credits: 20_000,
+            credits: 20000,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 0,
             schooling: 'Feuerwehr - Werkfeuerwehr-Ausbildung',
             shownSchooling: 'Werkfeuerwehr',
+            icon: 'truck',
         },
-        {
+        86: {
             caption: 'Turbolöscher',
             color: '#993422',
             coins: 30,
-            credits: 12_500,
+            credits: 12500,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 0,
             schooling: 'Feuerwehr - Werkfeuerwehr-Ausbildung',
             shownSchooling: 'Werkfeuerwehr',
+            icon: 'truck',
         },
-        {
+        87: {
             caption: 'TLF 4000',
             color: '#dd2222',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 3,
             wtank: 4000,
+            icon: 'truck-moving',
         },
-        {
+        88: {
             caption: 'KLF',
             color: '#660000',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 6,
             wtank: 500,
+            icon: 'truck-moving',
         },
-        {
+        89: {
             caption: 'MLF',
             color: '#550000',
             coins: 25,
-            credits: 5_000,
+            credits: 5000,
             minPersonnel: 1,
             maxPersonnel: 6,
-            wtank: 1_000,
+            wtank: 1000,
+            icon: 'truck-moving',
         },
-        {
+        90: {
             caption: 'HLF 10',
             color: '#330000',
             coins: 25,
-            credits: 20_000,
+            credits: 20000,
             minPersonnel: 1,
             maxPersonnel: 9,
-            wtank: 1_000,
+            wtank: 1000,
+            icon: 'truck-moving',
         },
-        {
+        91: {
             caption: 'Rettungshundefahrzeug',
             color: '#663300',
             coins: 25,
-            credits: 25_000,
+            credits: 25000,
             minPersonnel: 4,
             maxPersonnel: 5,
             schooling: 'Rettungsdienst - Rettungshundeführer (SEG)',
             shownSchooling: 'Rettungshundeführer (SEG)',
             special:
                 'Am Einsatzort sind ein "Rettungshundefahrzeug" und ein "Anh Hund" vom THW gleichwertig!',
+            icon: 'paw',
         },
-        {
+        92: {
             caption: 'Anh Hund',
             color: '#422629',
             coins: 0,
-            credits: 6_000,
+            credits: 6000,
             special:
                 'Muss von einem "MTW-OV" an den Einsatzort gezogen werden. Am Einsatzort sind ein "Anh Hund" und ein "Rettungshundefahrzeug" vom Rettungsdienst gleichwertig!',
+            icon: 'paw',
         },
-        {
+        93: {
             caption: 'MTW-OV',
             color: '#14165e',
             coins: 0,
-            credits: 19_000,
+            credits: 19000,
             minPersonnel: 4,
             maxPersonnel: 5,
             schooling: 'THW - Rettungshundeführer (THW)',
             shownSchooling: 'Rettungshundeführer (THW)',
+            icon: 'paw',
         },
-        {
+        94: {
             caption: 'DHuFüKw',
             color: '#339900',
             coins: 10,
-            credits: 15_000,
+            credits: 15000,
             minPersonnel: 1,
             maxPersonnel: 2,
             schooling: 'Polizei - Hundeführer (Schutzhund)',
             shownSchooling: 'Hundeführer (Schutzhund)',
+            icon: 'paw',
         },
-    ],
-    buildings: [
-        {
+        95: {
+            caption: 'Polizeimotorrad',
+            color: '#004400',
+            coins: 10,
+            credits: 3000,
+            minPersonnel: 1,
+            maxPersonnel: 1,
+            schooling: 'Polizei - Motorradstaffel',
+            shownSchooling: 'Motorradstaffel',
+            special:
+                'Das Polizeimotorrad kann im Spiel anstelle eines FuStW verwendet werden. Der Nachteil: Es kann keine Gefangenen transportieren und arbeitet nur halb so schnell wie ein FuStW.',
+            icon: 'motorcycle',
+        },
+    },
+    buildings: {
+        0: {
             caption: 'Feuerwache',
             color: '#bb0000',
             coins: 30,
@@ -1079,13 +1194,13 @@ export default {
             ],
             maxBuildingsFunction: (): number => 4_000,
         },
-        {
+        1: {
             caption: 'Feuerwehrschule',
             color: '#992222',
             coins: 50,
             credits: 500_000,
             extensions: new Array(3).fill({
-                caption: 'Klassenzimmer',
+                caption: 'Weiterer Klassenraum',
                 credits: 400_000,
                 coins: 40,
                 duration: '7 Tage',
@@ -1098,7 +1213,7 @@ export default {
             startPersonnel: 0,
             startVehicles: [],
         },
-        {
+        2: {
             caption: 'Rettungswache',
             color: '#ffa500',
             coins: 35,
@@ -1111,13 +1226,13 @@ export default {
             startPersonnel: 3,
             startVehicles: ['RTW'],
         },
-        {
+        3: {
             caption: 'Rettungsschule',
             color: '#ddc722',
             coins: 50,
             credits: 500_000,
             extensions: new Array(3).fill({
-                caption: 'Klassenzimmer',
+                caption: 'Weiterer Klassenraum',
                 credits: 400_000,
                 coins: 40,
                 duration: '7 Tage',
@@ -1130,7 +1245,7 @@ export default {
             startPersonnel: 0,
             startVehicles: [],
         },
-        {
+        4: {
             caption: 'Krankenhaus',
             color: '#bbe944',
             coins: 25,
@@ -1143,7 +1258,7 @@ export default {
                     duration: '7 Tage',
                 },
                 {
-                    caption: 'Allgemeine Chirugie',
+                    caption: 'Allgemeine Chirurgie',
                     credits: 10_000,
                     coins: 10,
                     duration: '7 Tage',
@@ -1199,7 +1314,7 @@ export default {
             startPersonnel: 0,
             startVehicles: [],
         },
-        {
+        5: {
             caption: 'Rettungshubschrauber-Station',
             color: '#e7ad2f',
             coins: 50,
@@ -1217,7 +1332,7 @@ export default {
                     ? 4
                     : Math.floor(buildingsAmountTotal / 25),
         },
-        {
+        6: {
             caption: 'Polizeiwache',
             color: '#007700',
             coins: 35,
@@ -1245,7 +1360,7 @@ export default {
             startVehicles: ['FuStW'],
             maxBuildingsFunction: (): number => 1_500,
         },
-        {
+        7: {
             caption: 'Leitstelle',
             color: '#24c3ae',
             coins: 0,
@@ -1260,13 +1375,13 @@ export default {
             maxBuildingsFunction: (buildingsAmountTotal: number): number =>
                 Math.floor(buildingsAmountTotal / 25) + 1,
         },
-        {
+        8: {
             caption: 'Polizeischule',
             color: '#225522',
             coins: 50,
             credits: 500_000,
             extensions: new Array(3).fill({
-                caption: 'Klassenzimmer',
+                caption: 'Weiterer Klassenraum',
                 credits: 400_000,
                 coins: 40,
                 duration: '7 Tage',
@@ -1279,20 +1394,20 @@ export default {
             startPersonnel: 0,
             startVehicles: [],
         },
-        {
+        9: {
             caption: 'THW-Ortsverein',
             color: '#000f76',
             coins: 35,
             credits: 200_000,
             extensions: [
                 {
-                    caption: 'Bergungsgruppe',
+                    caption: '1. Technischer Zug: Bergungsgruppe 2',
                     credits: 25_000,
                     coins: 5,
                     duration: '7 Tage',
                 },
                 {
-                    caption: 'Zugtrupp',
+                    caption: '1. Technischer Zug: Zugtrupp',
                     credits: 25_000,
                     coins: 5,
                     duration: '7 Tage',
@@ -1310,7 +1425,7 @@ export default {
                     duration: '7 Tage',
                 },
                 {
-                    caption: '2. Technischer Zug',
+                    caption: '2. Technischer Zug - Grundvorraussetzungen',
                     credits: 25_000,
                     coins: 5,
                     duration: '7 Tage',
@@ -1342,13 +1457,13 @@ export default {
             startPersonnel: 9,
             startVehicles: ['GKW'],
         },
-        {
+        10: {
             caption: 'THW Bundesschule',
             color: '#222d54',
             coins: 50,
             credits: 500_000,
             extensions: new Array(3).fill({
-                caption: 'Klassenzimmer',
+                caption: 'Weiterer Klassenraum',
                 credits: 400_000,
                 coins: 40,
                 duration: '7 Tage',
@@ -1361,20 +1476,20 @@ export default {
             startPersonnel: 0,
             startVehicles: [],
         },
-        {
+        11: {
             caption: 'Bereitschaftspolizei',
             color: '#227722',
             coins: 50,
             credits: 500_000,
             extensions: [
                 {
-                    caption: '2. Zug der 1. Hunderschaft',
+                    caption: '2. Zug der 1. Hundertschaft',
                     coins: 5,
                     credits: 25_000,
                     duration: '3 Tage',
                 },
                 {
-                    caption: '3. Zug der 1. Hunderschaft',
+                    caption: '3. Zug der 1. Hundertschaft',
                     coins: 5,
                     credits: 25_000,
                     duration: '3 Tage',
@@ -1430,7 +1545,7 @@ export default {
             startPersonnel: 0,
             startVehicles: [],
         },
-        {
+        12: {
             caption: 'Schnelleinsatzgruppe (SEG)',
             color: '#e05b00',
             coins: 30,
@@ -1468,7 +1583,7 @@ export default {
             startPersonnel: 0,
             startVehicles: ['KTW Typ B'],
         },
-        {
+        13: {
             caption: 'Polizeihubschrauberstation',
             color: '#148423',
             coins: 50,
@@ -1486,7 +1601,7 @@ export default {
                     ? 4
                     : Math.floor(buildingsAmountTotal / 25),
         },
-        {
+        14: {
             caption: 'Bereitstellungsraum',
             coins: 0,
             credits: 0,
@@ -1500,7 +1615,7 @@ export default {
             startVehicles: [],
             maxBuildingsFunction: (): number => 1,
         },
-        {
+        15: {
             caption: 'Wasserrettung',
             color: '#7fffd4',
             coins: 30,
@@ -1514,7 +1629,7 @@ export default {
             startPersonnel: 6,
             startVehicles: ['GW-Wasserrettung'],
         },
-        {
+        16: {
             caption: 'Verbandszellen',
             coins: 'x',
             credits: 100_000,
@@ -1532,7 +1647,7 @@ export default {
             startPersonnel: 0,
             startVehicles: [],
         },
-        {
+        17: {
             caption: 'Polizei-Sondereinheiten',
             color: '#1a7e23',
             coins: 40,
@@ -1576,7 +1691,7 @@ export default {
             startPersonnel: 0,
             startVehicles: [],
         },
-        {
+        18: {
             caption: 'Feuerwache (Kleinwache)',
             color: '#aa1111',
             coins: 25,
@@ -1615,7 +1730,7 @@ export default {
             ],
             maxBuildingsFunction: (): number => 4_000,
         },
-        {
+        19: {
             caption: 'Polizeiwache (Kleinwache)',
             color: '#116611',
             coins: 25,
@@ -1640,7 +1755,7 @@ export default {
             startVehicles: ['FuStW'],
             maxBuildingsFunction: (): number => 1_500,
         },
-        {
+        20: {
             caption: 'Rettungswache (Kleinwache)',
             color: '#eeb611',
             coins: 25,
@@ -1658,7 +1773,7 @@ export default {
             startPersonnel: 3,
             startVehicles: ['RTW'],
         },
-        {
+        21: {
             caption: 'Rettungshundestaffel',
             color: '#663300',
             coins: 50,
@@ -1671,7 +1786,7 @@ export default {
             startPersonnel: 10,
             startVehicles: ['Rettungshundefahrzeug'],
         },
-    ],
+    },
     buildingCategories: {
         Feuerwehr: {
             buildings: [0, 1, 18],
@@ -1731,7 +1846,7 @@ export default {
         },
         Polizei: {
             vehicles: {
-                'Funkstreifenwagen': [32],
+                'Funkstreifenwagen': [32, 95],
                 'Bereitschaftspolizei-Fahrzeuge': [35, 50, 51, 52, 72],
                 'Polizeihubschrauber': [61],
                 'SEK': [79, 80],
@@ -1849,6 +1964,10 @@ export default {
             {
                 caption: 'MEK',
                 duration: '7 Tage',
+            },
+            {
+                caption: 'Motorradstaffel',
+                duration: '3 Tage',
             },
             {
                 caption: 'Polizeihubschrauber',
@@ -2023,6 +2142,7 @@ export default {
         'Holzverarbeitung',
         'Motorsportanlage',
         'Tunnel',
+        'Klärwerk',
     ],
     only_alliance_missions: [41, 43, 59, 75, 99, 207, 221, 222, 256, 350],
     transfer_missions: [147],
