@@ -1,4 +1,1249 @@
+// import { Building } from 'typings/Building';
+
+const moduleRootFiles = require.context('../', true, MODULE_ROOT_I18N_FILES);
+const furtherFiles = require.context('./en_US/', true, /.*(\/index)?\.js(on)?/);
+const modules = {
+    appstore: {
+        save: 'Save',
+        reset: 'Reset',
+        noMapkit:
+            'This module does not work with the map type "Mapkit" due to limitations of Mapkit!',
+        dev:
+            'This module is currently still under development. Activating it can lead to incomplete and faulty functions!',
+        closeWarning: {
+            title: 'Unsaved changes',
+            text:
+                'You made changes in the AppStore that have not yet been saved. Reset them or save them to close the AppStore.',
+            close: 'Close message',
+        },
+    },
+    settings: {
+        name: 'Settings',
+        save: 'Save',
+        discard: 'Discard changes',
+        reset: 'Reset',
+        export: 'Export',
+        import: 'Import',
+        resetWarning: {
+            title: 'Reset the settings',
+            text:
+                'Do you really want to reset settings to their default values? This cannot be undone!',
+            close: 'Cancel',
+            total: 'All settings',
+            module: 'Only from this module',
+        },
+        closeWarning: {
+            title: 'Unsaved changes',
+            text:
+                'You have made changes in the settings that are not yet saved. Reset them, discard them or save them to close the settings.',
+            close: 'Close message',
+        },
+        changeList: {
+            true: 'On',
+            false: 'Off',
+        },
+    },
+} as { [moduleId: string]: { [key: string]: unknown } };
+moduleRootFiles
+    .keys()
+    .forEach(key => (modules[key.split('/')[2]] = moduleRootFiles(key)));
+
+const t = {} as { [key: string]: unknown };
+
+furtherFiles
+    .keys()
+    .forEach(
+        key => (t[key.split('/')[1].replace(/\..*$/, '')] = furtherFiles(key))
+    );
+
 export default {
-    settings: 'Settings',
+    modules,
+    ...t,
+    error: {
+        title: 'LSS Manager: Error',
+        msg:
+            'If this error occurs frequently, please report it to the LSSM team!',
+    },
+    warnings: {
+        version: {
+            title: 'Wrong LSS Manager version',
+            text:
+                'Dear user, unfortunately we had to discover that you do not have the latest version of LSS Manager. The latest version is {version}, but you have {curver} first. Please reload the game without cache (with Ctrl + F5, on Apple devices command + R), this should fix the bug. If the error persists, please report it to the team! If you use a wrong version we cannot guarantee the full functionality of the LSS-Manager.',
+            close: 'Close message and reload game (recommended)',
+            abort: 'Close message without reloading game',
+        },
+    },
+    globalSettings: {
+        name: 'General Settings',
+        labelInMenu: {
+            title: 'Label instead of icon in menu',
+            description:
+                'Displays a simple label in the navigation bar instead of the LSSM logo',
+        },
+    },
+    vehicles: {
+        0: {
+            caption: 'Type 1 fire engine',
+            color: '#cc0000',
+            coins: 25,
+            credits: 5_000,
+            minPersonnel: 1,
+            maxPersonnel: 6,
+            wtank: 0,
+        },
+        1: {
+            caption: 'Type 2 fire engine',
+            color: '#bb0000',
+            coins: 25,
+            credits: 5_000,
+            minPersonnel: 1,
+            maxPersonnel: 3,
+            wtank: 0,
+        },
+        2: {
+            caption: 'Platform truck',
+            color: '#d92626',
+            coins: 30,
+            credits: 10_000,
+            minPersonnel: 1,
+            maxPersonnel: 3,
+            wtank: 0,
+            special: 'Required once you have built 3 firehouses',
+        },
+        3: {
+            caption: 'Battalion chief unit',
+            color: '#d02525',
+            coins: 25,
+            credits: 10_000,
+            minPersonnel: 1,
+            maxPersonnel: 3,
+            wtank: 0,
+            special: 'Required once you have built 6 firehouses',
+        },
+        4: {
+            caption: 'Heavy rescue vehicle',
+            color: '#ad1f1f',
+            coins: 25,
+            credits: 12_180,
+            minPersonnel: 1,
+            maxPersonnel: 4,
+            wtank: 0,
+            special: 'Required once you have built 4 firehouses',
+        },
+        5: {
+            caption: 'ALS Ambulance',
+            color: '#9c1c1c',
+            coins: 25,
+            credits: 5_000,
+            minPersonnel: 1,
+            maxPersonnel: 3,
+            wtank: 0,
+        },
+        6: {
+            caption: 'Mobile air',
+            color: '#aa0000',
+            coins: 25,
+            credits: 11_860,
+            minPersonnel: 1,
+            maxPersonnel: 3,
+            wtank: 0,
+            special: 'Required once you have built 5 firehouses',
+        },
+        7: {
+            caption: 'Water Tanker',
+            color: '#990000',
+            coins: 25,
+            credits: 17_300,
+            minPersonnel: 1,
+            maxPersonnel: 3,
+            wtank: 3_000,
+            special: 'Required once you have built 7 firehouses',
+        },
+        8: {
+            caption: 'Utility unit',
+            color: '#880000',
+            coins: 25,
+            credits: 12_180,
+            minPersonnel: 1,
+            maxPersonnel: 4,
+            wtank: 0,
+            special: 'Required once you have built 4 firehouses',
+        },
+        9: {
+            caption: 'HazMat',
+            color: '#770000',
+            coins: 25,
+            credits: 19_200,
+            minPersonnel: 1,
+            maxPersonnel: 6,
+            wtank: 0,
+            special: 'Required once you have built 11 firehouses',
+            schooling: 'Fire Station - Hazmat',
+            shownSchooling: 'Hazmat',
+        },
+        10: {
+            caption: 'Patrol Car',
+            color: '#8b1818',
+            coins: 25,
+            credits: 5_000,
+            minPersonnel: 1,
+            maxPersonnel: 2,
+            wtank: 0,
+        },
+        11: {
+            caption: 'HEMS',
+            color: '#e61919',
+            coins: 30,
+            credits: 300_000,
+            minPersonnel: 1,
+            maxPersonnel: 1,
+            wtank: 0,
+        },
+        12: {
+            caption: 'Mobile command vehicle',
+            color: '#791515',
+            coins: 25,
+            credits: 25_500,
+            minPersonnel: 1,
+            maxPersonnel: 6,
+            wtank: 0,
+            schooling: 'Fire Station - Mobile command',
+            shownSchooling: 'Mobile command',
+            special: 'Required once you have built 13 firehouses',
+        },
+        13: {
+            caption: 'Quint',
+            color: '#dc1818',
+            coins: 25,
+            credits: 19_000,
+            minPersonnel: 1,
+            maxPersonnel: 6,
+            wtank: 0,
+            special:
+                'To purchase with credits it requires the rank: Captain, <br>Lower ranked members can purchase the vehicle for 25 Coins. <br>Quint acts as a Platform Truck and a Fire Truck.',
+        },
+        14: {
+            caption: 'Police helicopter',
+            color: '#ca1616',
+            coins: 30,
+            credits: 300_000,
+            minPersonnel: 1,
+            maxPersonnel: 2,
+            wtank: 0,
+            schooling: 'Police - Police Aviation',
+            shownSchooling: 'Police Aviation',
+        },
+        15: {
+            caption: 'Fly-Car',
+            color: '#b81414',
+            coins: 20,
+            credits: 4_000,
+            minPersonnel: 1,
+            maxPersonnel: 2,
+            wtank: 0,
+        },
+        16: {
+            caption: 'SWAT Armoured Vehicle',
+            color: '#a51212',
+            coins: 25,
+            credits: 17_300,
+            minPersonnel: 6,
+            maxPersonnel: 6,
+            wtank: 0,
+            schooling: 'Police - Swat',
+            shownSchooling: 'SWAT',
+            special: 'Required once you have built 8 police stations',
+        },
+        17: {
+            caption: 'ARFF Crash Tender',
+            color: '#cc2222',
+            coins: 25,
+            credits: 20_000,
+            minPersonnel: 1,
+            maxPersonnel: 3,
+            wtank: 0,
+            schooling: 'Fire Station - ARFF',
+            shownSchooling: 'ARFF',
+        },
+        18: {
+            caption: 'Rescue Engine',
+            color: '#bb2222',
+            coins: 25,
+            credits: 19_000,
+            minPersonnel: 1,
+            maxPersonnel: 6,
+            wtank: 0,
+            special:
+                'To purchase with credits it requires the rank: Captain, <br>Lower ranked members can purchase the vehicle for 25 Coins.<br> The Rescue Engine acts as a Heavy Rescue and a Fire Engine.',
+        },
+        19: {
+            caption: 'K-9 Unit',
+            color: '#aa2222',
+            coins: 25,
+            credits: 7_000,
+            minPersonnel: 1,
+            maxPersonnel: 2,
+            wtank: 0,
+            schooling: 'Police - K-9',
+            shownSchooling: 'K-9',
+            special: 'Required from 6 Police stations',
+        },
+        20: {
+            caption: 'Mass Casualty Unit',
+            color: '#992222',
+            coins: 25,
+            credits: 25_000,
+            minPersonnel: 6,
+            maxPersonnel: 6,
+            wtank: 0,
+            special:
+                'The MCU can be alarmed in two different modes:<br>1. for lightly injured persons who do not need an emergency physician. Then 7 patients can be treated and removed simultaneously in the MCU.<br>2.For severely injured persons who need an emergency physician. Then 3 patients can be needled simultaneously in the MCU.<br>You can buy 1 Mass Casualty Vehicle for every 20 ambulance stations (respectively 15 with premium account).',
+        },
+        21: {
+            caption: 'Heavy Rescue + Boat',
+            color: '#882222',
+            coins: 25,
+            credits: 19_000,
+            minPersonnel: 1,
+            maxPersonnel: 4,
+            wtank: 0,
+            schooling: 'Fire Station - Swift water rescue',
+            shownSchooling: 'Swift water rescue',
+            special:
+                'Combines the Heavy Rescue Vehicle/Utility Vehicle with an integrated boat.<br> No need for a boat trailer.',
+        },
+        22: {
+            caption: 'Boat Trailer',
+            color: '#772222',
+            coins: 12,
+            credits: 6_000,
+            minPersonnel: 0,
+            maxPersonnel: 0,
+            wtank: 0,
+            special:
+                'Accessory boat that can be towed with the Utility Unit. Requires Swift Water Rescue trained personnel to operate.<br> The boat trailer cannot be assigned personnel, <br>either assign your trained personnel to the towing vehicle, or assign them to a separate unit en route.',
+        },
+        23: {
+            caption: 'Police Motorcycle',
+            color: '#662222',
+            coins: 18,
+            credits: 2_500,
+            minPersonnel: 1,
+            maxPersonnel: 1,
+            wtank: 0,
+            schooling: 'Police - Motor Officer',
+            shownSchooling: 'Motor Officer',
+        },
+        24: {
+            caption: 'Large Fireboat',
+            color: '#552222',
+            coins: 35,
+            credits: 35_000,
+            minPersonnel: 2,
+            maxPersonnel: 7,
+            wtank: 0,
+            schooling: 'Fire Station - Ocean Navigation',
+            shownSchooling: 'Ocean Navigation',
+        },
+        25: {
+            caption: 'Large Rescue Boat',
+            color: '#442222',
+            coins: 35,
+            credits: 35_000,
+            minPersonnel: 2,
+            maxPersonnel: 6,
+            wtank: 0,
+            schooling: 'Fire Station - Ocean Navigation',
+            shownSchooling: 'Ocean Navigation',
+        },
+        26: {
+            caption: 'SWAT SUV',
+            color: '#332222',
+            coins: 23,
+            credits: 7_000,
+            minPersonnel: 2,
+            maxPersonnel: 4,
+            wtank: 0,
+            special: 'Required once you have built 8 police stations',
+            schooling: 'Police - SWAT',
+            shownSchooling: 'SWAT',
+        },
+        27: {
+            caption: 'BLS Ambulance',
+            color: '#681212',
+            coins: 25,
+            credits: 5_000,
+            minPersonnel: 1,
+            maxPersonnel: 2,
+            wtank: 0,
+        },
+        28: {
+            caption: 'EMS Rescue',
+            color: '#f59f00',
+            coins: 25,
+            credits: 12_180,
+            minPersonnel: 1,
+            maxPersonnel: 2,
+            special:
+                'No missions require a EMS Rescue at this time, however it is able to treat patients. And Works as a Heavy Rescue',
+        },
+        29: {
+            caption: 'EMS Chief',
+            color: '#e09200',
+            coins: 20,
+            credits: 10_000,
+            minPersonnel: 1,
+            maxPersonnel: 2,
+            special: 'Required once you have built 10 Rescue stations',
+        },
+        30: {
+            caption: 'Type 3 engine',
+            color: '#440000',
+            coins: 19,
+            credits: 19_000,
+            minPersonnel: 3,
+            maxPersonnel: 5,
+            wtank: 500,
+        },
+        31: {
+            caption: 'Type 5 engine',
+            color: '#ddaa3c',
+            coins: 8,
+            credits: 8_000,
+            minPersonnel: 1,
+            maxPersonnel: 3,
+        },
+        32: {
+            caption: 'Type 7 engine',
+            color: '#005500',
+            coins: 5,
+            credits: 5_000,
+            minPersonnel: 1,
+            maxPersonnel: 2,
+            wtank: 50,
+        },
+        33: {
+            caption: 'Pumper Tanker',
+            color: '#570f0f',
+            coins: 19,
+            credits: 19_000,
+            minPersonnel: 1,
+            maxPersonnel: 6,
+            wtank: 1500,
+        },
+        34: {
+            caption: 'Crew Carrier',
+            color: '#bf2222',
+            coins: 10,
+            credits: 10_000,
+            minPersonnel: 1,
+            maxPersonnel: 12,
+            wtank: 0,
+        },
+        35: {
+            caption: 'Water drop helicopter',
+            color: '#288f28',
+            coins: 130,
+            credits: 300_000,
+            minPersonnel: 2,
+            maxPersonnel: 5,
+            wtank: 2000,
+            schooling: 'Fire Station - Airborne firefighting',
+            shownSchooling: 'Airborne firefighting',
+        },
+        36: {
+            caption: 'Air Tanker',
+            color: '#450c0c',
+            coins: 50,
+            credits: 1_000_000,
+            minPersonnel: 2,
+            maxPersonnel: 5,
+            wtank: 7200,
+            schooling: 'Fire Station - Airborne firefighting',
+            shownSchooling: 'Airborne firefighting',
+        },
+        37: {
+            caption: 'Heavy air tanker',
+            color: '#220000',
+            coins: 65,
+            credits: 1_500_000,
+            minPersonnel: 2,
+            maxPersonnel: 5,
+            wtank: 12000,
+            schooling: 'Fire Station - Airborne firefighting',
+            shownSchooling: 'Airborne firefighting',
+        },
+        38: {
+            caption: 'Type 4 engine',
+            color: '#ffb61a',
+            coins: 10,
+            credits: 10_000,
+            minPersonnel: 3,
+            maxPersonnel: 5,
+            wtank: 750,
+        },
+        39: {
+            caption: 'Type 6 engine',
+            color: '#00138f',
+            coins: 5,
+            credits: 5_000,
+            minPersonnel: 1,
+            maxPersonnel: 2,
+            wtank: 150,
+        },
+        40: {
+            caption: 'Dozer Trailer',
+            color: '#0016a3',
+            coins: 15,
+            credits: 20_000,
+            minPersonnel: 0,
+            maxPersonnel: 0,
+            schooling: 'Fire Station - Heavy Machinery Operating',
+            shownSchooling: 'Heavy Machinery Operating',
+            special:
+                "Dozer Trailer that can be towed with the Crew cap semi. Requires Heavy Machinery Operating and Truck Driver's License trained personnel to operate.<br> The dozer trailer cannot be assigned personnel, <br>either assign your trained personnel to the towing vehicle, or assign them to a separate unit en route.",
+        },
+        41: {
+            caption: 'Crew cap semi',
+            color: '#0018b8',
+            coins: 12,
+            credits: 5_000,
+            minPersonnel: 1,
+            maxPersonnel: 3,
+            schooling: "Fire Station - Truck Driver's License",
+            shownSchooling: "Truck Driver's License",
+        },
+        42: {
+            caption: 'FBI Unit',
+            color: '#001bcc',
+            coins: 15,
+            credits: 10_000,
+            minPersonnel: 2,
+            maxPersonnel: 4,
+        },
+        43: {
+            caption: 'FBI Investigation Wagon',
+            color: '#001ee0',
+            coins: 15,
+            credits: 10_000,
+            minPersonnel: 2,
+            maxPersonnel: 4,
+        },
+        44: {
+            caption: 'FBI Mobile Command Center',
+            color: '#0021f5',
+            coins: 25,
+            credits: 25_000,
+            minPersonnel: 1,
+            maxPersonnel: 4,
+            schooling: 'Police - FBI Mobile Center Commmander',
+            shownSchooling: 'FBI Mobile Center Commmander',
+        },
+        45: {
+            caption: 'FBI Bomb Technician Vehicle',
+            color: '#0a2bff',
+            coins: 35,
+            credits: 35_500,
+            minPersonnel: 1,
+            maxPersonnel: 2,
+            schooling: 'Police - FBI Bomb Technician',
+            shownSchooling: 'FBI Bomb Technician',
+        },
+        46: {
+            caption: 'FBI Surveillance Drone',
+            color: '#f04242',
+            coins: 25,
+            credits: 25_000,
+            minPersonnel: 1,
+            maxPersonnel: 1,
+            wtank: 0,
+            schooling: 'Police - FBI Drone Operator',
+            shownSchooling: 'FBI Drone Operator',
+        },
+        47: {
+            caption: 'Sheriff Unit',
+            color: '#ad0e0e',
+            coins: 20,
+            credits: 15_000,
+            minPersonnel: 1,
+            maxPersonnel: 2,
+            schooling: 'Police - Sheriff',
+            shownSchooling: 'Sheriff',
+        },
+    },
+    buildings: {
+        0: {
+            caption: 'Fire station',
+            color: '#bb0000',
+            coins: 30,
+            credits: 100_000,
+            extensions: [
+                {
+                    caption: 'Ambulance extension',
+                    credits: 100_000,
+                    coins: 20,
+                    duration: '7 Days',
+                },
+                {
+                    caption: 'Airport extension',
+                    credits: 100_000,
+                    coins: 20,
+                    duration: '7 Days',
+                },
+                {
+                    caption: 'Forestry Expansion',
+                    credits: 50_000,
+                    coins: 13,
+                    duration: '7 Days',
+                },
+                {
+                    caption: 'Water rescue expansion',
+                    credits: 100_000,
+                    coins: 20,
+                    duration: '7 Days',
+                },
+            ],
+            levelcost: ['1. 10.000', '2. 50.000', '3.-39. 100.000'],
+            maxBuildings: '4.000 together with small fire stations',
+            maxLevel: 39,
+            special:
+                'From the 24th fire station onwards, the cost of building a new fire station increases according to the following formula: <code>100.000+200.000*LOG<sub>2</sub>(Number of existing fire stations − 22)</code>. The Coins price remains constant!',
+            startPersonnel: 10,
+            startVehicles: ['Type 1 fire engine', 'Type 2 fire engine'],
+            maxBuildingsFunction: (): number => 4_000,
+        },
+        1: {
+            caption: 'Dispatch Center',
+            color: '#24c3ae',
+            coins: 0,
+            credits: 0,
+            extensions: [],
+            levelcost: [],
+            maxBuildings: 'All 25 buildings one control center',
+            maxLevel: 0,
+            special: 'The control center is the administrative center.',
+            startPersonnel: 0,
+            startVehicles: [],
+            maxBuildingsFunction: (buildingsAmountTotal: number): number =>
+                Math.floor(buildingsAmountTotal / 25) + 1,
+        },
+        2: {
+            caption: 'Hospital',
+            color: '#bbe944',
+            coins: 25,
+            credits: 200_000,
+            extensions: [
+                {
+                    caption: 'General Internal',
+                    credits: 10_000,
+                    coins: 10,
+                    duration: '7 Days',
+                },
+                {
+                    caption: 'General Surgeon',
+                    credits: 10_000,
+                    coins: 10,
+                    duration: '7 Days',
+                },
+                {
+                    caption: 'Gynecology',
+                    credits: 70_000,
+                    coins: 15,
+                    duration: '7 Days',
+                },
+                {
+                    caption: 'Urology',
+                    credits: 70_000,
+                    coins: 15,
+                    duration: '7 Days',
+                },
+                {
+                    caption: 'Traumatology',
+                    credits: 70_000,
+                    coins: 15,
+                    duration: '7 Days',
+                },
+                {
+                    caption: 'Neurology',
+                    credits: 70_000,
+                    coins: 15,
+                    duration: '7 Days',
+                },
+                {
+                    caption: 'Neurosurgery',
+                    credits: 70_000,
+                    coins: 15,
+                    duration: '7 Days',
+                },
+                {
+                    caption: 'Cardiology',
+                    credits: 70_000,
+                    coins: 15,
+                    duration: '7 Days',
+                },
+                {
+                    caption: 'Cardiac Surgery',
+                    credits: 70_000,
+                    coins: 15,
+                    duration: '7 Days',
+                },
+            ],
+            levelcost: ['1.-20. 19.000 Credits / 11 Coins'],
+            maxBuildings: 'No limit',
+            maxLevel: 20,
+            special:
+                'Finance ministers and admins can (expand) association hospitals with the help of credits from the association treasury.',
+            startPersonnel: 0,
+            startVehicles: [],
+        },
+        3: {
+            caption: 'Ambulance station',
+            color: '#ffa500',
+            coins: 35,
+            credits: 200_000,
+            extensions: [],
+            levelcost: ['1. 10.000', '2. 50.000', '3.-39. 100.000'],
+            maxBuildings: 'No limit',
+            maxLevel: 39,
+            special: '',
+            startPersonnel: 3,
+            startVehicles: ['ALS Ambulance'],
+        },
+        4: {
+            caption: 'Fire academy',
+            color: '#992222',
+            coins: 50,
+            credits: 500_000,
+            extensions: [
+                ...new Array(3).fill({
+                    caption: 'Additional classroom',
+                    credits: 400_000,
+                    coins: 40,
+                    duration: '7 Days',
+                }),
+            ],
+            levelcost: [],
+            maxBuildings: 'No limit',
+            maxLevel: 0,
+            special:
+                "Finance ministers and admins can (expand) fire department schools with the help of credits from the association's treasury.Training course masters and admins can start training courses at association fire- brigade schools.",
+            startPersonnel: 0,
+            startVehicles: [],
+        },
+        5: {
+            caption: 'Police station',
+            color: '#007700',
+            coins: 35,
+            credits: 100_000,
+            extensions: [
+                {
+                    caption: 'Prison cell',
+                    credits: 25_000,
+                    coins: 5,
+                    duration: '7 Days',
+                },
+                ...new Array(9).fill({
+                    caption: 'Additional cell',
+                    credits: 25_000,
+                    coins: 5,
+                    duration: '7 Days',
+                }),
+            ],
+            levelcost: ['1. 10.000', '2. 50.000', '3.-39. 100.000'],
+            maxBuildings: '1.500 together with small police stations',
+            maxLevel: 39,
+            special:
+                'From the 24th police station onwards, the costs for the new construction of a police station increase according to the following formula: <code>100.000+200.000*LOG<sub>2</sub>(Number of existing police stations − 22)</code>. The Coins price remains constant!',
+            startPersonnel: 2,
+            startVehicles: ['Patrol car'],
+            maxBuildingsFunction: (): number => 1_500,
+        },
+        6: {
+            caption: 'Helicopter station',
+            color: '#e7ad2f',
+            coins: 50,
+            credits: 1_000_000,
+            extensions: [],
+            levelcost: [],
+            maxBuildings: 'see specials',
+            maxLevel: 0,
+            special:
+                'Up to the 125th building (of all types) a total of max. 4 landing sites can be built. After that the number increases by 1 every 25 buildings (starting at the 125th).',
+            startPersonnel: 0,
+            startVehicles: [],
+            maxBuildingsFunction: (buildingsAmountTotal: number): number =>
+                buildingsAmountTotal < 125
+                    ? 4
+                    : Math.floor(buildingsAmountTotal / 25),
+        },
+        7: {
+            caption: 'Police academy',
+            color: '#225522',
+            coins: 50,
+            credits: 500_000,
+            extensions: [
+                ...new Array(3).fill({
+                    caption: 'Additional classroom',
+                    credits: 400_000,
+                    coins: 40,
+                    duration: '7 Days',
+                }),
+            ],
+            levelcost: [],
+            maxBuildings: 'No limit',
+            maxLevel: 0,
+            special:
+                "Finance ministers and admins can (expand) association police schools with the help of credits from the association's treasury.Training course masters and admins can start training courses at association police schools.",
+            startPersonnel: 0,
+            startVehicles: [],
+        },
+        8: {
+            caption: 'Police Aviation',
+            color: '#148423',
+            coins: 50,
+            credits: 1_000_000,
+            extensions: [],
+            levelcost: ['1. 1.000.000 Credits / 50 Coins'],
+            maxBuildings: 'see specials',
+            maxLevel: 1,
+            special:
+                'Up to 2 landing sites can be built per station (expansion stages). Up to the 125th building (of all types) a total of max. 4 landing sites can be built. After that the number increases by 1 every 25 buildings (starting at the 125th).',
+            startPersonnel: 3,
+            startVehicles: [],
+            maxBuildingsFunction: (buildingsAmountTotal: number): number =>
+                buildingsAmountTotal < 125
+                    ? 4
+                    : Math.floor(buildingsAmountTotal / 25),
+        },
+        9: {
+            caption: 'Staging area',
+            coins: 0,
+            credits: 0,
+            extensions: [],
+            levelcost: [],
+            maxBuildings: 4,
+            maxLevel: 0,
+            special:
+                'You can station as many of your own vehicles as you like at a staging area, members of the association can use the room. A staging area remains for 24 hours, but you can reset it to 24 hours at any time.With Premium Account you can have 8 stating areas at the same time',
+            startPersonnel: 0,
+            startVehicles: [],
+            maxBuildingsFunction: (): number => 4,
+        },
+        10: {
+            caption: 'Prison Cells',
+            coins: 'x',
+            credits: 100_000,
+            extensions: [
+                {
+                    caption: 'Prison cell',
+                    credits: 25_000,
+                    coins: 5,
+                    duration: '7 Days',
+                },
+                ...new Array(9).fill({
+                    caption: 'Additional cell',
+                    credits: 25_000,
+                    coins: 5,
+                    duration: '7 Days',
+                }),
+            ],
+            levelcost: [],
+            maxBuildings: 'No limit',
+            maxLevel: 0,
+            special:
+                "This building can only be built and developed by admins and finance ministers with credits from the association's treasury.The built Prison Cells are available to all members of the association.",
+            startPersonnel: 0,
+            startVehicles: [],
+        },
+        11: {
+            caption: 'Fire Boat Dock',
+            color: '#663300',
+            coins: 35,
+            credits: 500_000,
+            extensions: [],
+            levelcost: ['1. 10.000', '2. 50.000', '3.-16. 100.000'],
+            maxBuildings: 'No limit',
+            maxLevel: 16,
+            special: '',
+            startPersonnel: 2,
+            startVehicles: [''],
+        },
+        12: {
+            caption: 'Rescue Boat Dock',
+            color: '#d3d62a',
+            coins: 35,
+            credits: 500_000,
+            extensions: [],
+            levelcost: ['1. 10.000', '2. 50.000', '3.-16. 100.000'],
+            maxBuildings: 'No limit',
+            maxLevel: 16,
+            special: '',
+            startPersonnel: 2,
+            startVehicles: [''],
+        },
+        13: {
+            caption: 'Fire station (Small station) ',
+            color: '#aa1111',
+            coins: 25,
+            credits: 50_000,
+            extensions: [],
+            levelcost: [
+                '1. 10.000',
+                '2. 50.000',
+                '3.-5. 100.000',
+                'Conversion to normal guard: difference price to normal guard',
+            ],
+            maxBuildings: '4.000 together with fire stations',
+            maxLevel: 5,
+            special:
+                'From the 24th fire station onwards, the cost of building a new fire station increases according to the following formula: <code>(100.000+200.000*LOG<sub>2</sub>(Number of existing fire stations − 22)) / 2</code>. max. 1 Million Credits. The Coins price remains constant!',
+            startPersonnel: 10,
+            startVehicles: ['Type 1 fire engine', 'Type 2 fire engine'],
+            maxBuildingsFunction: (): number => 4_000,
+        },
+        14: {
+            caption: 'Clinic',
+            color: '#e2e53b',
+            coins: 25,
+            credits: 100_000,
+            extensions: [
+                {
+                    caption: 'General Internal',
+                    credits: 10_000,
+                    coins: 10,
+                    duration: '7 Days',
+                },
+            ],
+            levelcost: ['1. 10.000', '2. 50.000', '3.-5. 100.000'],
+            maxBuildings: 'No limit',
+            maxLevel: 5,
+            special: '',
+            startPersonnel: 0,
+            startVehicles: ['Non. You can buy max. 2 Vehicles'],
+        },
+        15: {
+            caption: 'Police station (Small station)',
+            color: '#116611',
+            coins: 25,
+            credits: 50_000,
+            extensions: [],
+            levelcost: [
+                '1. 10.000',
+                '2. 50.000',
+                '3.-4. 100.000',
+                'Conversion to normal guard: difference price to normal guard',
+            ],
+            maxBuildings: '1.500 mit Polizeiwachen zusammen',
+            maxLevel: 4,
+            special:
+                'From the 24th police station onwards, the costs for the new construction of a police station are calculated according to the following formula: <code>(100.000+200.000*LOG<sub>2</sub>(Number of existing police stations − 22)) / 2</code>. The Coins price remains constant!',
+            startPersonnel: 2,
+            startVehicles: ['Patrol Car'],
+            maxBuildingsFunction: (): number => 1_500,
+        },
+        16: {
+            caption: 'Ambulance station (Small station)',
+            color: '#eeb611',
+            coins: 25,
+            credits: 100_000,
+            extensions: [],
+            levelcost: [
+                '1. 10.000',
+                '2. 50.000',
+                '3.-5. 100.000',
+                'Conversion to normal guard: difference price to normal guard',
+            ],
+            maxBuildings: 'No limit',
+            maxLevel: 5,
+            special: '',
+            startPersonnel: 3,
+            startVehicles: ['ALS Ambulance'],
+        },
+        17: {
+            caption: 'Firefighting plane station',
+            color: '#db7918',
+            coins: 65,
+            credits: 1_500_000,
+            extensions: [],
+            levelcost: ['1. 1.500.000'],
+            maxBuildings: 'No limit',
+            maxLevel: 2,
+            special: '',
+            startPersonnel: 2,
+            startVehicles: ['Water drop helicopter'],
+        },
+        18: {
+            caption: 'Federal Police Station',
+            color: '#268926',
+            coins: 50,
+            credits: 5_000_000,
+            extensions: [],
+            levelcost: ['1. 10.000', '2. 50.000', '3.-20. 100.000'],
+            maxBuildings: 'No limit',
+            maxLevel: 20,
+            special: '',
+            startPersonnel: 10,
+            startVehicles: ['FBI Unit'],
+        },
+    },
+    buildingCategories: {
+        'Fire Department': {
+            buildings: [0, 4, 11, 13, 17],
+            color: '#ff2d2d',
+        },
+        'Rescue Stations': {
+            buildings: [3, 6, 12, 16],
+            color: '#ffa500',
+        },
+        'Police Stations': {
+            buildings: [5, 7, 8, 15, 18],
+            color: '#00ac00',
+        },
+        'Other': {
+            buildings: [2, 9, 14],
+            color: '#02a18c',
+        },
+    },
+    vehicleCategories: {
+        'Firefighters': {
+            vehicles: {
+                'Fire trucks': [0, 1, 13, 18],
+                'WaterTanker': [7],
+                'Special vehicles': [2, 3, 4, 6, 8, 9, 12],
+                'Wildland vehicles': [30, 31, 32, 33, 34, 38, 39, 40, 41],
+                'Airport Vehicles': [17],
+                'Boats': [21, 22, 24],
+                'Fire Aviation': [35, 36, 37],
+            },
+            color: '#ff2d2d',
+        },
+        'Rescue Vehicles': {
+            vehicles: {
+                'Ambulances': [5, 27],
+                'HEMS': [11],
+                'First Responder': [15],
+                'Rescue Boat': [25],
+                'Other EMS Vehicle': [20, 28, 29],
+            },
+            color: '#ffa500',
+        },
+        'Police Vehicles': {
+            vehicles: {
+                'Patrol Car': [10],
+                'SWAT': [16, 26],
+                'Police Motorcycle': [23],
+                'Police helicopter': [14],
+                'K-9 Unit': [19],
+                'FBI': [42, 43, 44, 45, 46],
+                'Sheriff': [47],
+            },
+            color: '#00ac00',
+        },
+    },
+    small_buildings: {
+        0: 13,
+        3: 16,
+        5: 15,
+    },
+    vehicleBuildings: [0, 3, 5, 6, 8, 9, 11, 12, 13, 14, 15, 16, 17, 18],
+    cellBuildings: [5],
+    cellExtensions: [
+        '5_0',
+        '5_1',
+        '5_2',
+        '5_3',
+        '5_4',
+        '5_5',
+        '5_6',
+        '5_7',
+        '5_8',
+        '5_9',
+    ],
+    bedBuildings: [2, 14],
+    schoolBuildings: [4, 7],
+    dispatchCenterBuildings: [1],
+    schoolings: {
+        'Fire Station': [
+            {
+                caption: 'Hazmat',
+                duration: '3 Days',
+            },
+            {
+                caption: 'Mobile command',
+                duration: '5 Days',
+            },
+            {
+                caption: 'ARFF-Training',
+                duration: '3 Days',
+            },
+            {
+                caption: 'Swift water rescue',
+                duration: '4 Days',
+            },
+            {
+                caption: 'Ocean Navigation',
+                duration: '5 Days',
+            },
+            {
+                caption: 'Airborne firefighting',
+                duration: '5 Days',
+            },
+            {
+                caption: 'Heavy Machinery Operating',
+                duration: '3 Days',
+            },
+            {
+                caption: "Truck Driver's License",
+                duration: '2 Days',
+            },
+        ],
+        'Police': [
+            {
+                caption: 'Police Aviation',
+                duration: '7 Days',
+            },
+            {
+                caption: 'SWAT',
+                duration: '5 Days',
+            },
+            {
+                caption: 'K-9',
+                duration: '5 Days',
+            },
+            {
+                caption: 'Police Motorcycle',
+                duration: '3 Days',
+            },
+            {
+                caption: 'FBI Mobile Center Commander ',
+                duration: '7 Days',
+            },
+            {
+                caption: 'FBI Bomb Technician',
+                duration: '5 Days',
+            },
+            {
+                caption: 'FBI Drone Operator',
+                duration: '5 Days',
+            },
+            {
+                caption: 'Sheriff',
+                duration: '5 Days',
+            },
+        ],
+    },
+    amount: 'Quantity',
+    search: 'Search',
+    alliance: 'Alliance',
+    premiumNotice:
+        'This feature extends a premium feature of the game and is therefore only available for players with a Missionchief game premium account!',
+    credits: 'Credits',
+    close: 'Close',
+    fullscreen: {
+        expand: 'Activate full screen mode',
+        compress: 'Disable full screen mode',
+    },
+    hideTitle: 'Show heading | Hide heading',
+    vehicle: 'Cars | Car | Cars',
+    building: 'Buildings',
+    station: 'Stations | Station | Stations',
+    distance: 'Distance | Distances',
+    vehicleType: 'Vehicle type',
+    fmsReal2Show: {
+        1: 1,
+        2: 2,
+        3: 3,
+        4: 4,
+        5: 5,
+        6: 6,
+        7: 7,
+        9: 9,
+    },
+    buildingIcons: [
+        'fire',
+        'rss',
+        'hospital',
+        'clinic-medical',
+        'graduation-cap',
+        'shield-alt',
+        'helicopter',
+        'graduation-cap',
+        'helicopter',
+        'building',
+        '',
+        'ship',
+        'ship',
+        'fire',
+        'stethoscope',
+        'shield-alt',
+        'clinic-medical',
+        'plane',
+        'shield-alt',
+    ],
+    pois: [
+        'Park',
+        'Lake',
+        'Hospital',
+        'Forest',
+        'Bus stop',
+        'Tram stop',
+        'Train station (regional traffic)',
+        'Train station (regional traffic and long-distance travel)',
+        'Goods station',
+        'Supermarket (small)',
+        'Supermarket (big)',
+        'Gas station',
+        'School',
+        'Museum',
+        'Mall',
+        'Car workshop',
+        'Highway exit',
+        'Christmas market',
+        'Storehouse',
+        'Discotheque',
+        'Stadium',
+        'Farm',
+        'Office Building',
+        'Swimming bath',
+        'Railroad Crossing',
+        'Theater',
+        'Fairground',
+        'River',
+        'Small Airport (Runway)',
+        'Large Airport (Runway)',
+        'Airport Terminal',
+        'Bank',
+        'Warehouse',
+        'Bridge',
+        'Fast Food Restaurant',
+        'Cargo port',
+        'Recycling Centre',
+        'High Rise',
+        'Cruise ship dock',
+        'Marina',
+        'Rail Crossing',
+        'Tunnel',
+        'Cold Storage Warehouse',
+        'Power Plant',
+        'Factory',
+        'Scrap yard',
+        'Subway station',
+        'Small chemical storage tank',
+        'Large chemical storage tank',
+        'Hotel',
+        'Bar',
+        'Landfill site',
+        'Parking Garage',
+        'Propane depot',
+        'Large Ferry Dock',
+        'Forest Cabin',
+        'Campsite',
+        'Bus Terminal',
+        'Pier',
+        'Fault line',
+        'Carpentry Workshop',
+        'Playground',
+    ],
+    only_alliance_missions: [41, 61, 62, 112],
     transfer_missions: [246],
 };
