@@ -1,5 +1,5 @@
 import { BuildingMarker, RadioMessage } from '../../../typings/Ingame';
-import { Vehicle } from '../../../typings/Vehicle';
+import { InternalVehicle, Vehicle } from '../../../typings/Vehicle';
 import { PointTuple } from 'leaflet';
 import { Building } from '../../../typings/Building';
 import { ModuleMainFunction } from 'typings/Module';
@@ -8,8 +8,10 @@ export default (async (LSSM, MODULE_ID) => {
     await LSSM.$store.dispatch('api/registerBuildingsUsage', true);
     await LSSM.$store.dispatch('api/registerVehiclesUsage', true);
 
-    const vehicleTypes = Object.values(LSSM.$t('vehicles')).map(
-        type => type.caption
+    const vehicleTypes = Object.fromEntries(
+        Object.entries(
+            LSSM.$t('vehicles') as { [id: number]: InternalVehicle }
+        ).map(([id, { caption }]) => [id, caption])
     );
 
     await LSSM.$store.dispatch('addStyle', {
@@ -57,9 +59,9 @@ export default (async (LSSM, MODULE_ID) => {
 
         if (building) {
             if (
-                Object.values(LSSM.$t('vehicleBuildings')).indexOf(
+                Object.values(LSSM.$t('vehicleBuildings')).includes(
                     building.building_type
-                ) >= 0
+                )
             ) {
                 data += `<br><i class="fa fa-parking"></i>&nbsp;${building.level +
                     1}&nbsp;<i class="fa fa-car"></i>&nbsp;${
@@ -67,7 +69,11 @@ export default (async (LSSM, MODULE_ID) => {
                 }&nbsp;<i class="fa fa-users"></i>&nbsp;${
                     building.personal_count
                 }`;
-                if (building.building_type === 6)
+                if (
+                    Object.values(LSSM.$t('cellBuildings')).includes(
+                        building.building_type
+                    )
+                )
                     data += `&nbsp;<i class="fa fa-border-all"></i>&nbsp;${
                         building.extensions.filter(x => x.available).length
                     }&nbsp;(${building.extensions.length})`;
@@ -87,16 +93,16 @@ export default (async (LSSM, MODULE_ID) => {
                 });
                 data += `</table>`;
             } else if (
-                Object.values(LSSM.$t('bedBuildings')).indexOf(
+                Object.values(LSSM.$t('bedBuildings')).includes(
                     building.building_type
-                ) >= 0
+                )
             ) {
                 data += `<br><i class="fa fa-procedures"></i>&nbsp;${building.level +
                     10}`;
             } else if (
-                Object.values(LSSM.$t('schoolBuildings')).indexOf(
+                Object.values(LSSM.$t('schoolBuildings')).includes(
                     building.building_type
-                ) >= 0
+                )
             ) {
                 data += `<br><i class="fa fa-chalkboard-teacher"></i>&nbsp;${building
                     .extensions.length + 1}`;
