@@ -10,7 +10,8 @@
                 <grid-board
                     :id="
                         $store.getters.nodeAttribute(
-                            'dispatchcenter-view_board'
+                            'dispatchcenter-view_board',
+                            true
                         )
                     "
                 >
@@ -83,7 +84,8 @@
                                 <grid-board
                                     :id="
                                         $store.getters.nodeAttribute(
-                                            `dispatchcenter-view_board-${column.building}`
+                                            `dispatchcenter-view_board-${column.building}`,
+                                            true
                                         )
                                     "
                                 >
@@ -140,7 +142,8 @@
                             :maxHeight="3"
                             :id="
                                 $store.getters.nodeAttribute(
-                                    'dispatchcenter-view_board-selection'
+                                    'dispatchcenter-view_board-selection',
+                                    true
                                 )
                             "
                             :width="buildingSelection.width"
@@ -279,7 +282,8 @@
                 <grid-board
                     :id="
                         $store.getters.nodeAttribute(
-                            'dispatchcenter-view_manage'
+                            'dispatchcenter-view_manage',
+                            true
                         )
                     "
                 >
@@ -359,7 +363,7 @@
 
 <script lang="ts">
 import Vue from 'vue';
-import { Building } from 'typings/Building';
+import { Building, InternalBuilding } from 'typings/Building';
 import {
     DispatchcenterView,
     DispatchcenterViewComputed,
@@ -399,7 +403,9 @@ export default Vue.extend<
             ).DashItem,
     },
     data() {
-        const buildingTypes = Object.values(this.$t('buildings'));
+        const buildingTypes = this.$t('buildings') as {
+            [id: number]: InternalBuilding;
+        };
         const dispatchCenterBuildings = Object.values(
             this.$t('dispatchCenterBuildings')
         );
