@@ -7,6 +7,8 @@ export default (async (LSSM, MODULE_ID, $m: $m) => {
     const poi_types = Object.values(LSSM.$t('pois')) as string[];
     poi_types.sort();
 
+    await LSSM.$store.dispatch('api/registerSettings');
+
     const style = await (async () => {
         const predef = await LSSM.$store.dispatch('settings/getSetting', {
             moduleId: MODULE_ID,
@@ -95,6 +97,10 @@ export default (async (LSSM, MODULE_ID, $m: $m) => {
     const poiHighlightedClass = LSSM.$store.getters.nodeAttribute(
         'poi-highlighted'
     );
+    const poiSettingsWrapperId = LSSM.$store.getters.nodeAttribute(
+        'poi-settings',
+        true
+    );
 
     await LSSM.$store.dispatch('addStyles', [
         {
@@ -138,7 +144,8 @@ export default (async (LSSM, MODULE_ID, $m: $m) => {
             .map(poi => `#map .poi[caption="${poi}"]`)
             .join(',');
         const extraStyleId = LSSM.$store.getters.nodeAttribute(
-            'poi-hider-style'
+            'poi-hider-style',
+            true
         );
         document.getElementById(extraStyleId)?.remove();
         const style = document.createElement('style');
@@ -147,11 +154,15 @@ export default (async (LSSM, MODULE_ID, $m: $m) => {
         document.body.append(style);
     };
     refresh_shown_pois();
-
+    const paddingLeftPOI = [3, 4].includes(
+        LSSM.$store.state.api.settings.design_mode
+    )
+        ? '25px'
+        : '1ch';
     const observer = new MutationObserver(mutations => {
         mutations.forEach(mutation => {
             const form = (mutation.target as HTMLElement).querySelector(
-                '#new_mission_position'
+                '#new_mission_position[action="/mission_positions"]'
             );
             if (!form) {
                 isPOIWindow = false;
@@ -162,7 +173,8 @@ export default (async (LSSM, MODULE_ID, $m: $m) => {
                 );
                 return;
             }
-            if (isPOIWindow) return;
+            if (isPOIWindow && document.getElementById(poiSettingsWrapperId))
+                return;
             isPOIWindow = true;
 
             colorMarkers(
@@ -178,7 +190,8 @@ export default (async (LSSM, MODULE_ID, $m: $m) => {
                 )
             );
             const settingsWrapper = document.createElement('div');
-            settingsWrapper.style.paddingLeft = '1ch';
+            settingsWrapper.style.paddingLeft = paddingLeftPOI;
+            settingsWrapper.id = poiSettingsWrapperId;
             form.append(settingsWrapper);
             settingsWrapper.append(
                 ...['all', 'none', ...poi_types].map(poi => {
