@@ -15,9 +15,9 @@ export default (LSSM: Vue, $m: $m): void => {
     const vehicle = (LSSM.$store.state.api.vehicles as Vehicle[]).find(
         v => v.id === vehicleId
     );
-    const vehicleTypes = Object.values(
-        LSSM.$t('vehicles')
-    ) as InternalVehicle[];
+    const vehicleTypes = LSSM.$t('vehicles') as {
+        [id: number]: InternalVehicle;
+    };
     if (vehicleId < 0 || !vehicle) return;
 
     const fittingRows = [] as HTMLTableRowElement[];
@@ -33,7 +33,8 @@ export default (LSSM: Vue, $m: $m): void => {
     });
 
     const toggleId = LSSM.$store.getters.nodeAttribute(
-        'toggle-fitting-personnel'
+        'toggle-fitting-personnel',
+        true
     );
 
     const settingsBar = document.createElement('form');
