@@ -1,24 +1,36 @@
 import moment from 'moment';
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { $m } from 'typings/Module';
 
 moment.locale(BUILD_LANG);
 
-export default (LSSM: Vue, $m: $m): void => {
+export default (LSSM: Vue, yellowBorder: number, redBorder: boolean): void => {
     LSSM.$store.commit('useFontAwesome');
 
     const generationDate = moment(
         document
-            .querySelector('#missionH1')
-            ?.getAttribute('data-original-title')
-            ?.replace(/^.*?:/, '')
-            .trim(),
-        $m('generationDate.inputFormat').toString()
+            .getElementById('mission_general_info')
+            ?.getAttribute('data-generation-time')
     );
 
-    const generationDateNode = document.createElement('i');
+    const general_info = document.querySelector(
+        '#mission_general_info > small'
+    );
+    if (!general_info) return;
+
+    const generationDateNode = document.createElement('span');
     generationDateNode.innerHTML = `&nbsp;|&nbsp;<i class="fas fa-history"></i>&nbsp;${generationDate.fromNow()} (${generationDate.calendar()})`;
     document
         .querySelector('#mission_general_info > small')
-        ?.appendChild(generationDateNode);
+        ?.append(generationDateNode);
+
+    generationDateNode.style.paddingLeft = '2px';
+    generationDateNode.style.paddingRight = '2px';
+
+    if (yellowBorder && moment().diff(generationDate, 'hours') >= yellowBorder)
+        generationDateNode.style.border = 'yellow 1px solid';
+
+    if (
+        redBorder &&
+        (generationDate.hours() <= 3 || generationDate.day() !== moment().day())
+    )
+        generationDateNode.style.border = 'red 1px solid';
 };
