@@ -81,6 +81,10 @@ module.exports = {
             grtw0: 'GRTW (7 Patienten - ohne Notarzt)',
             grtw1: 'GRTW (3 Patienten - inkl. Notarzt)',
             fustw: 'Funkstreifenwagen',
+            police_motorcycle: 'Polizeimotorrad',
+            fustw_or_police_motorcycle:
+                'Funkstreifenwagen oder Polizeimotorrad',
+            k9: 'DHuFüKw',
             lebefkw: 'Leichter Befehlskraftwagen (leBefKw)',
             fukw: 'FüKw (Führungskraftwagen)',
             grukw: 'GruKw (Gruppenkraftwagen)',
@@ -117,9 +121,6 @@ module.exports = {
             rescue_dogs: 'Anhänger Hundetransport oder Rettungshundefahrzeug',
         },
     },
-    generationDate: {
-        inputFormat: 'DD. MMMM, HH:mm Uhr',
-    },
     enhancedMissingVehicles: {
         vehicle: 'Fahrzeugtyp',
         missing: 'Am Einsatzort fehlend',
@@ -134,7 +135,7 @@ module.exports = {
             [/^ELW 1$/]: [3, 34, 78],
             [/^ELW 2$/]: [34, 78],
             [/^Flugfeldlöschfahrzeuge$/]: [75],
-            [/^FuStW$/]: [32],
+            [/^FuStW$/]: [32, 95],
             [/^FwK$/]: [57],
             [/^FüKw$/]: [51],
             [/^GefKw$/]: [52],
@@ -147,6 +148,7 @@ module.exports = {
             [/^GW-Öl$/]: [10, 49],
             [/^Gerätekraftwagen \(GKW\)$/]: [39],
             [/^GruKw$/]: [50],
+            [/^GW-San$/]: [60],
             [/^LKW Kipper \(LKW K 9\)$/]: [42],
             [/^Löschfahrzeuge? \(LF\)$/]: [
                 0,
@@ -181,6 +183,7 @@ module.exports = {
             [/^RTW oder KTW oder KTW-B$/]: [28, 38, 58],
             [/^Rüstwagen oder HLF$/]: [4, 30, 90],
             [/^SEK-Fahrzeuge$/]: [79, 80],
+            [/^DHuFüKw$/]: [94],
             [/^Schlauchwagen \(GW-L2 Wasser, SW 1000, SW 2000 oder Ähnliches\)$/]: [
                 11,
                 13,
@@ -199,107 +202,6 @@ module.exports = {
         },
     },
     tailoredTabs: {
-        defaultTabs: [
-            {
-                name: 'LF',
-                vehicleTypes: [
-                    0,
-                    1,
-                    6,
-                    7,
-                    8,
-                    9,
-                    17,
-                    18,
-                    19,
-                    20,
-                    21,
-                    22,
-                    23,
-                    24,
-                    25,
-                    26,
-                    30,
-                    37,
-                    87,
-                    88,
-                    89,
-                    90,
-                ],
-            },
-            {
-                name: 'FW andere',
-                vehicleTypes: [
-                    2,
-                    3,
-                    4,
-                    5,
-                    10,
-                    11,
-                    12,
-                    13,
-                    14,
-                    15,
-                    16,
-                    27,
-                    33,
-                    34,
-                    36,
-                    46,
-                    47,
-                    48,
-                    49,
-                    53,
-                    54,
-                    57,
-                    62,
-                    71,
-                    75,
-                    76,
-                    77,
-                    78,
-                    83,
-                    84,
-                    85,
-                    86,
-                ],
-            },
-            {
-                name: 'Rettung',
-                vehicleTypes: [28, 29, 31, 38, 55, 56, 58, 59, 60, 73, 74, 91],
-            },
-            {
-                name: 'Polizei',
-                vehicleTypes: [32, 61],
-            },
-            {
-                name: 'Bereitschaftspolizei',
-                vehicleTypes: [35, 50, 51, 52, 72, 79, 80, 81, 82],
-            },
-            {
-                name: 'THW',
-                vehicleTypes: [
-                    39,
-                    40,
-                    41,
-                    42,
-                    43,
-                    44,
-                    45,
-                    65,
-                    66,
-                    67,
-                    68,
-                    69,
-                    92,
-                    93,
-                ],
-            },
-            {
-                name: 'Wasser',
-                vehicleTypes: [63, 64, 70],
-            },
-        ],
         allTab: 'Alle',
         occupiedTab: 'Weiteralarmieren',
     },
